% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{available_datasets}
\alias{available_datasets}
\title{Get meta information about available datasets.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 7 columns.
}
\usage{
available_datasets
}
\value{
A tibble of metadata about covidregionaldata.
}
\description{
Lists available datasets including: function names,
data returned, subregion level / raw data sources, alongside
package information. It can also be used to help guide adding
data sources to the package.
}
\keyword{datasets}
