% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu.R
\name{check_alternate_data_source}
\alias{check_alternate_data_source}
\title{Filter JHU data for target country}
\usage{
check_alternate_data_source(target_country)
}
\arguments{
\item{target_country}{A string with the country of interest}
}
\value{
A data frame of COVID cases for the target country
}
\description{
Downloads daily covid data from JHU (via get_jhu_data)
and subsets it for a given target country
}
