% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data_sources}
\alias{check_data_sources}
\title{Check data sources}
\usage{
check_data_sources(
  countries = c("afghanistan", "belgium", "brazil", "canada", "colombia", "germany",
    "india", "italy", "russia", "uk", "usa"),
  worldwide = TRUE
)
}
\arguments{
\item{countries}{names of countries as in the "country.R" file name to check}

\item{worldwide}{report on worldwide datasets - ECDC and WHO, interventions, linelist}
}
\value{
A tibble of latest dates for all sources and data checks for sub-national data
}
\description{
Check that data are up to date and returning correctly
}
