% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_level}
\alias{check_level}
\title{Checks a given level is supported}
\usage{
check_level(level, supported_levels)
}
\arguments{
\item{level}{A character string indicating the current level.}

\item{supported_levels}{A character vector of supported levels}
}
\description{
Checks a given level is supported
}
\concept{utility}
