% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national.R
\name{clean_regional.crd_ecdc_1}
\alias{clean_regional.crd_ecdc_1}
\title{ECDC Specific Country Level Data Cleaning}
\usage{
\method{clean_regional}{crd_ecdc_1}(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Clean downloaded ECDC data
}
\examples{
\dontrun{
ecdc <- new_covidregionaldata("ecdc")
ecdc <- download_regional(ecdc)
clean_regional(ecdc)$clean
}
}
\author{
Sam Abbott @seabbs

D. van Muijen @dmuijen

Kath Sherratt @kathsherratt
}
