% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy.R
\name{clean_regional.crd_italy_1}
\alias{clean_regional.crd_italy_1}
\title{Italy Specific State Level Data Cleaning}
\usage{
\method{clean_regional}{crd_italy_1}(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Italy Specific State Level Data Cleaning
}
\examples{
\dontrun{
italy <- new_covidregionaldata("italy")
italy <- download_regional(italy)
clean_regional(italy)$clean
}
}
\author{
Sam Abbott
}
