% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico.R
\name{clean_regional.crd_mexico_2}
\alias{clean_regional.crd_mexico_2}
\title{Mexico Specific Municipality Level Data Cleaning}
\usage{
\method{clean_regional}{crd_mexico_2}(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Mexico Specific Municipality Level Data Cleaning
}
\examples{
\dontrun{
mexico <- new_covidregionaldata("mexico", level = "2")
mexico <- download_regional(mexico)
clean_regional(mexico)$clean
}
}
\author{
Sam Abbott
}
