% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{clean_regional.crd_who_1}
\alias{clean_regional.crd_who_1}
\title{WHO Specific Country Level Data Cleaning}
\usage{
\method{clean_regional}{crd_who_1}(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Clean downloaded WHO data
}
\examples{
\dontrun{
who <- new_covidregionaldata("who")
who <- download_regional(who)
clean_regional(who)$clean
}
}
\author{
Sam Abbott @seabbs

Kath Sherratt @kathsherratt
}
