% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-defaults.R
\name{clean_regional.default}
\alias{clean_regional.default}
\title{Default region specific data cleaning}
\usage{
\method{clean_regional}{default}(region, verbose, ...)
}
\description{
Default regional cleaning method. Highlights that no custom
cleaning is happening.
}
\author{
Sam Abbott
}
