% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-methods.R
\name{clean_regional}
\alias{clean_regional}
\title{Region specific data cleaning}
\usage{
clean_regional(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Region specific data cleaning. See the underlying
regional methods for details. Designed to be called after
\code{download_regional}
}
\author{
Sam Abbott
}
