% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{colombia_codes}
\alias{colombia_codes}
\title{Region Codes for Colombia Dataset.}
\format{
An object of class \code{data.frame} with 33 rows and 2 columns.
}
\usage{
colombia_codes
}
\value{
A tibble of region codes and related information.
}
\description{
The region codes for Colombia
}
\keyword{datasets}
