% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_covid19R_format}
\alias{convert_to_covid19R_format}
\title{Convert data to Covid19R package data standard}
\usage{
convert_to_covid19R_format(data)
}
\arguments{
\item{data}{A data frame / tibble}
}
\value{
A data frame in the Covid19R standard
}
\description{
Converts wide format (time series) data into long format to meet the Covid19R package standard
}
