% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csv_reader}
\alias{csv_reader}
\title{Custom CSV reading function}
\usage{
csv_reader(file, verbose = FALSE, ...)
}
\arguments{
\item{file}{A URL or filepath to a CSV}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{extra parameters to be passed to vroom::vroom}
}
\value{
A data table
}
\description{
Checks for use of memoise and then uses vroom::vroom.
}
