% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csv_readr}
\alias{csv_readr}
\title{Custom CSV reading function}
\usage{
csv_readr(file, ...)
}
\arguments{
\item{file}{A URL or filepath to a CSV}

\item{...}{extra parameters to be passed to readr::read_csv}
}
\value{
A data table
}
\description{
Checks for use of memoise and then uses readr::read_csv, which appears
more robust in loading some streams
}
