% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-methods.R
\name{DataClass}
\alias{DataClass}
\title{R6 Class containing non-country specific methods}
\description{
Acts as parent class for individual country objects,
allowing them to access general methods.
}
\details{
All countries have shared methods for extracting region codes,
downloading, processing, and returning data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{country}}{name of country to fetch data for}

\item{\code{data}}{data frame for requested region}

\item{\code{supported_levels}}{A list of supported levels.}

\item{\code{supported_region_names}}{A list of region names in order of level.}

\item{\code{supported_region_codes}}{A list of region codes in order of level.}

\item{\code{region_name}}{string Name for the codes column, e.g. 'iso_3166_2'}

\item{\code{code_name}}{string Name for the codes column, e.g. 'iso_3166_2'}

\item{\code{codes_lookup}}{string or tibble Region codes for the target country}

\item{\code{data_url}}{List of named links to raw data. The first, and
sometimes only entry, should be named main}

\item{\code{source_data_cols}}{existing columns within the raw data}

\item{\code{level}}{target region level}

\item{\code{totals}}{Boolean. If TRUE, returns totalled data per region
up to today's date.}

\item{\code{localise}}{Boolean. Should region names be localised.}

\item{\code{verbose}}{Boolean. Display information at various stages.}

\item{\code{steps}}{Boolean. Keep data from each processing step.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_region_codes}{\code{DataClass$set_region_codes()}}
\item \href{#method-new}{\code{DataClass$new()}}
\item \href{#method-download}{\code{DataClass$download()}}
\item \href{#method-clean}{\code{DataClass$clean()}}
\item \href{#method-process}{\code{DataClass$process()}}
\item \href{#method-return}{\code{DataClass$return()}}
\item \href{#method-summary}{\code{DataClass$summary()}}
\item \href{#method-clone}{\code{DataClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_region_codes"></a>}}
\if{latex}{\out{\hypertarget{method-set_region_codes}{}}}
\subsection{Method \code{set_region_codes()}}{
Place holder for custom country specific function to load
region codes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$set_region_codes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the country
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The args passed by \link{initialise_dataclass}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
Download raw data using the list
of supplied data_urls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$download()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\if{latex}{\out{\hypertarget{method-clean}{}}}
\subsection{Method \code{clean()}}{
Clean data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process"></a>}}
\if{latex}{\out{\hypertarget{method-process}{}}}
\subsection{Method \code{process()}}{
Processes data.
Dynamically works for level 1 and level 2 regions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$process()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return"></a>}}
\if{latex}{\out{\hypertarget{method-return}{}}}
\subsection{Method \code{return()}}{
Optional region specific return changes.
Designed to be called after \code{process}. For most datasets a
custom method should not be needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$return()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Class summary information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a single row summary tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
