% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-methods.R
\name{check_country_avaliable}
\alias{check_country_avaliable}
\title{Check country and regions are avaliable and set up country class}
\usage{
check_country_avaliable(
  country = character(),
  level = level,
  totals = totals,
  localise = localise,
  verbose = verbose,
  steps = steps,
  ...
)
}
\arguments{
\item{country}{A character string specifying the country to get data from.
Not case dependent. Name should be the English name. For a list of
options see \code{available_datasets}.}

\item{level}{A character string indicating the target administrative level
of the data with the default being "1". Currently supported options are
level 1 ("1) and level 2 ("2"). See \code{available_datasets} for supported
options by dataset.}

\item{totals}{Logical, defaults to FALSE. If TRUE, returns totalled
data per region up to today's date. If FALSE, returns the full dataset
stratified by date and region.}

\item{localise}{Logical, defaults to TRUE. Should region names be localised.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{steps}{Logical, defaults to FALSE. Should all processing and cleaning
steps be kept and output in a list.}

\item{...}{additional arguments to pass to Country classes.}
}
\value{
The target countries specific object if avaliable, e.g. \code{\link[=Italy]{Italy()}}
}
\description{
Check data for the requested country and region
is avaliable and return an initialised region class for that country.
}
