% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_excel}
\alias{download_excel}
\title{Download Excel Documents}
\usage{
download_excel(url, archive, verbose = FALSE, transpose = TRUE, ...)
}
\arguments{
\item{url}{Character string containing the full URL to the Excel document.}

\item{archive}{Character string naming the file name to assign in the
temporary directory.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{transpose}{Logical, should the read in data be transposed}

\item{...}{Additional parameters to pass to \code{read_excel()}.}
}
\value{
A \code{data.frame}.
}
\description{
Download Excel Documents
}
\concept{utility}
