% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy.R
\name{download_regional.crd_italy_1}
\alias{download_regional.crd_italy_1}
\title{Download Italian Regional Daily COVID-19 Count Data}
\usage{
\method{download_regional}{crd_italy_1}(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Extracts daily COVID-19 data for Italy, stratified by .
Data available at .
}
\examples{
\dontrun{
italy <- new_covidregionaldata("italy")
italy <- download_regional(italy)
italy$raw
}
}
\author{
Sam Abbott
}
