% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico.R
\name{download_regional.crd_mexico_1}
\alias{download_regional.crd_mexico_1}
\title{Download Mexican Regional Daily COVID-19 Count Data - Estado}
\usage{
\method{download_regional}{crd_mexico_1}(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Extracts daily COVID-19 data for Mexico, stratified by Estados.
Data available at \url{https://datos.covid-19.conacyt.mx/#DownZCSV}.
}
\examples{
\dontrun{
mexico <- new_covidregionaldata("mexico")
mexico <- download_regional(mexico)
mexico$raw
}
}
\author{
Sam Abbott
}
