% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{download_regional.crd_who_1}
\alias{download_regional.crd_who_1}
\title{Download WHO Daily COVID-19 Count Data}
\usage{
\method{download_regional}{crd_who_1}(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Downloads WHO Covid-19 data from the opendata
portal: https://opendata.ecdc.europa.eu/covid19/casedistribution/csv
}
\examples{
\dontrun{
who <- new_covidregionaldata("who")
who <- download_data(who)
who$raw
}
}
\author{
Sam Abbott
}
