% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-defaults.R
\name{download_regional.default}
\alias{download_regional.default}
\title{Default regional data downloading}
\usage{
\method{download_regional}{default}(region, verbose, ...)
}
\description{
Default download method. Causes an error if called.
}
\author{
Sam Abbott
}
