% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-methods.R
\name{download_regional}
\alias{download_regional}
\title{Region specific data downloading}
\usage{
download_regional(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Region specific raw data download function. See
region specific methods for details.
}
\author{
Sam Abbott
}
