% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_data.R
\name{Ecdc}
\alias{Ecdc}
\title{R6 Class containing specific attributes and methods for ECDC dataset}
\description{
Country specific information for downloading, cleaning
and procesing covid-19 region data for ECDC
}
\details{
Inherits from \code{dataClass}
}
\author{
Sam Abbott
}
\section{Super class}{
\code{\link[covidregionaldata:Generic methods]{covidregionaldata::Generic methods}} -> \code{ecdc}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level_1_region}}{the level 1 region name.}

\item{\code{level_2_region}}{the level 2 region name.}

\item{\code{data_url}}{link to raw data}

\item{\code{source_data_cols}}{existing columns within the raw data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clean_regional}{\code{Ecdc$clean_regional()}}
\item \href{#method-return_regional}{\code{Ecdc$return_regional()}}
\item \href{#method-new}{\code{Ecdc$new()}}
\item \href{#method-clone}{\code{Ecdc$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="Generic methods" data-id="download_data">}\href{../../covidregionaldata/html/Generic methods.html#method-download_data}{\code{covidregionaldata::Generic methods$download_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="Generic methods" data-id="get_region_codes">}\href{../../covidregionaldata/html/Generic methods.html#method-get_region_codes}{\code{covidregionaldata::Generic methods$get_region_codes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="Generic methods" data-id="process_regional">}\href{../../covidregionaldata/html/Generic methods.html#method-process_regional}{\code{covidregionaldata::Generic methods$process_regional()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_regional"></a>}}
\if{latex}{\out{\hypertarget{method-clean_regional}{}}}
\subsection{Method \code{clean_regional()}}{
ECDC specific state level data cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ecdc$clean_regional(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{pass additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_regional"></a>}}
\if{latex}{\out{\hypertarget{method-return_regional}{}}}
\subsection{Method \code{return_regional()}}{
Specific return settings for the ECDC dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ecdc$return_regional(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{pass additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Set up the country class with attributes set to input parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ecdc$new(level, totals, localise, verbose, steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The region level for the data}

\item{\code{totals}}{Boolean. If TRUE, returns totalled data per region up to today's date.}

\item{\code{localise}}{Boolean. Should region names be localised.}

\item{\code{verbose}}{Boolean. Display information at various stages.}

\item{\code{steps}}{Boolean. Keep data from each processing step.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ecdc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
