% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{expect_columns_contain_data}
\alias{expect_columns_contain_data}
\title{Test that cleaned columns contain data and are the correct types}
\usage{
expect_columns_contain_data(cntry_obj)
}
\arguments{
\item{cntry_obj}{The DataClass object (R6Class) object to perform checks on.}
}
\description{
Expect data has cleaned columns. Inherited by child
classes so tests run through each class.
}
