% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{fill_empty_dates_with_na}
\alias{fill_empty_dates_with_na}
\title{Add rows of NAs for dates where a region does not have any data}
\usage{
fill_empty_dates_with_na(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A tibble with rows of NAs added.
}
\description{
There are points, particularly early during data collection,
where data was not collected for all regions. This function finds dates
which have data for some regions, but not all, and adds rows of NAs for the
missing regions. This is mainly for reasons of completeness.
}
\seealso{
Compulsory processing functions
\code{\link{add_extra_na_cols}()},
\code{\link{calculate_columns_from_existing_data}()},
\code{\link{complete_cumulative_columns}()}
}
\concept{compulsory_processing}
