% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecdc_cases.R
\name{format_ecdc_data}
\alias{format_ecdc_data}
\title{Format ECDC data for use with EpiNow analysis}
\usage{
format_ecdc_data(data = NULL)
}
\arguments{
\item{data}{data.frame returned by get_ecdc_cases function}
}
\value{
data.frame with formatted ecdc case data by country
}
\description{
Format ECDC data for use with EpiNow analysis
}
\examples{
format_ecdc_data(get_ecdc_cases(countries = "France"))
}
\author{
Joel Hellewell
}
