% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_process_options}
\alias{format_process_options}
\title{Constructs list of options for processing}
\usage{
format_process_options(process_options)
}
\arguments{
\item{process_options}{list, additional arguments to control what
functions are called during processing. For some datasets setting these
to FALSE may cause errors, but for others improve speed.}
}
\description{
Contructs processing options including user arguments.
}
