% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{france_codes}
\alias{france_codes}
\title{Region Codes for France Dataset.}
\format{
An object of class \code{data.frame} with 104 rows and 5 columns.
}
\usage{
france_codes
}
\value{
A tibble of region codes and related information.
}
\description{
The region codes for France
}
\keyword{datasets}
