% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-methods.R
\name{general_init}
\alias{general_init}
\title{initialize function used by all \code{Country} class objects.}
\usage{
general_init(
  self,
  level = "1",
  totals = FALSE,
  localise = TRUE,
  verbose = TRUE,
  steps = FALSE
)
}
\arguments{
\item{self}{The specific class object to attach values}

\item{level}{The region level for the data}

\item{totals}{Boolean. If TRUE, returns totalled data per region up to
today's date.}

\item{localise}{Boolean. Should region names be localised.}

\item{verbose}{Boolean. Display information at various stages.}

\item{steps}{Boolean. Keep data from each processing step.}
}
\description{
Set up the country class with attributes set to input
parameters. Should only be called by a \code{Country} class object.
}
