% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{get_authority_lookup_table}
\alias{get_authority_lookup_table}
\title{Lookup table for local authority structure for the UK}
\usage{
get_authority_lookup_table()
}
\value{
A tibble of UK local authorities
}
\description{
Gets data from \url{https://opendata.arcgis.com/datasets/72e57d3ab1054e169c55afff3c9c1aa4_0.csv}
and then uses this to create a table of authorities and their corresponding higher level regions
}
