% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_datasets.R
\name{get_available_datasets}
\alias{get_available_datasets}
\title{Get the countries data is available for}
\usage{
get_available_datasets(type)
}
\arguments{
\item{type}{A character vector indiciating the the types of data to
return. Current options include "national" (which are datasets at the
national level which inherit from\code{CountryDataClass}) and
"regional" (which are datasets at the regional level which inherit
directly from \code{DataClass}).}
}
\value{
A list of available countries and the region level
data is available for
}
\description{
Show what countries have what level data available.
The function searches the environment for R6 class objects and
extracts the country name and what level it has from the object.
}
\examples{
get_available_datasets()
}
