% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_belgium_regional_cases.R
\name{get_belgium_regional_cases_only_level_1}
\alias{get_belgium_regional_cases_only_level_1}
\title{Fetch daily COVID cases with only Admin Level 1 region (region) for Belgium}
\usage{
get_belgium_regional_cases_only_level_1()
}
\value{
A data.frame of COVID cases by admin level 1 region in Belgium, ready to be used by get_regional_covid_data()
}
\description{
Fetches daily COVID data from Sciensano, the Belgian Institute for Health.
Data is available at https://epistat.wiv-isp.be/covid/
selects and sanitises the relevant columns
}
