% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/belgium.R
\name{get_belgium_regional_cases_with_level_2}
\alias{get_belgium_regional_cases_with_level_2}
\title{Fetch daily COVID cases including Admin Level 2 region (Province) for Belgium}
\usage{
get_belgium_regional_cases_with_level_2()
}
\value{
A data.frame of COVID cases by admin level 2 region in Belgium, ready to be used by get_regional_covid_data()
}
\description{
Fetches daily COVID data from Sciensano, the Belgian Institute for Health.
Data is available at https://epistat.wiv-isp.be/covid/
selects and sanitises the relevant columns. Also includes Level 1 region
}
