% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_brazil_level_2_codes}
\alias{get_brazil_level_2_codes}
\title{Brazilian level 2 codes (not available currently)}
\usage{
get_brazil_level_2_codes()
}
\description{
Brazilian level 2 codes (not available currently)
}
