% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_regional_cases.R
\name{get_brazil_regional_cases_only_level_1}
\alias{get_brazil_regional_cases_only_level_1}
\title{Brazil regional cases and deaths stratifying by Admin Level 1 region (state)}
\usage{
get_brazil_regional_cases_only_level_1()
}
\description{
Extracts regional case and death counts for Brazil, stratified by state.
Data available on Github, curated by Wesley Cota: DOI 10.1590/SciELOPreprints.362
\href{https://github.com/wcota/covid19br}{Source}
}
