% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_regional_cases.R
\name{get_brazil_regional_cases_with_level_2}
\alias{get_brazil_regional_cases_with_level_2}
\title{Brazil regional cases and deaths stratiftinf by Admin Level 2 region (city)}
\usage{
get_brazil_regional_cases_with_level_2()
}
\description{
Extracts regional case and death counts for Brazil, stratified by city
Data available on Github, curated by Wesley Cota: DOI 10.1590/SciELOPreprints.362
\href{https://github.com/wcota/covid19br}{Source}
}
