% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada.R
\name{get_canada_regional_cases}
\alias{get_canada_regional_cases}
\title{Helper function to fetch daily COVID cases by province for Canada and do dataset-specific processing before passing to get_regional_covid_data()}
\usage{
get_canada_regional_cases()
}
\value{
A data.frame of COVID cases by province in Canada, ready to be used by get_regional_covid_data()
}
\description{
Fetches daily COVID cases, and deaths by province collated by Provincial Canadian Health Authorities
Data is available at https://www.canada.ca/en/public-health/services/diseases/2019-novel-coronavirus-infection.html
selects the relevant columns, sanitises various columns and gets the daily counts from cumulative columns.
}
