% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso_codes.R
\name{get_colombia_iso_codes}
\alias{get_colombia_iso_codes}
\title{Colombia ISO codes (NULL - they're in the raw data already)}
\usage{
get_colombia_iso_codes()
}
\description{
Colombia ISO codes (NULL - they're in the raw data already)
}
