% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colombia.R
\name{get_colombia_regional_cases}
\alias{get_colombia_regional_cases}
\title{Get Colombia daily cases}
\usage{
get_colombia_regional_cases()
}
\value{
A dataframe of case, death and testing counts in Colombian regions
}
\description{
Fetches COVID case, death, and test counts by region
Sourced from Ideas Col: https://github.com/ideascol/covid19
}
\examples{


\dontrun{

regions <- rnaturalearth::ne_states("Colombia", returnclass = "sf")
data <- get_colombia_regional_cases() \%>\%
          dplyr::filter(date == max(date))
regions_with_data <- dplyr::left_join(regions, data, by = "iso_3166_2")
regions_with_data \%>\%
  ggplot2::ggplot() +
  ggplot2::geom_sf(ggplot2::aes(fill = cases))
}

}
