% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cumulative_from_daily}
\alias{get_cumulative_from_daily}
\title{Cumulative counts from daily counts.}
\usage{
get_cumulative_from_daily(column)
}
\arguments{
\item{column}{A vector of numeric data (e.g. a data column) which corresponds to daily counts of a process. Can contain NA}
}
\value{
A vector of numeric data which corresponds to cumulative counts
}
\description{
Gets cumulative cases/deaths etc. from data which is in daily forms. Similar to cumsum() but deals with NAs by treating them as 0.
}
