% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_daily_from_cumulative}
\alias{get_daily_from_cumulative}
\title{Daily counts from data that is in cumulative form.}
\usage{
get_daily_from_cumulative(column)
}
\arguments{
\item{column}{A vector of numeric data (e.g. a data column) which corresponds to cumulative counts of a process}
}
\value{
A vector of numeric data which corresponds to daily counts
}
\description{
Gets daily counts from data column that is in cumulative form.
}
