% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_france_level_2_codes}
\alias{get_france_level_2_codes}
\title{France level 2 codes (included in original function)}
\usage{
get_france_level_2_codes()
}
\description{
France level 2 codes (included in original function)
}
