% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_france_region_codes}
\alias{get_france_region_codes}
\title{France region codes (NULL - they're in the raw data already)}
\usage{
get_france_region_codes()
}
\description{
France region codes (NULL - they're in the raw data already)
}
