% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_france_regional_cases.R
\name{get_france_regional_cases}
\alias{get_france_regional_cases}
\title{French Regional Daily Case Counts}
\usage{
get_france_regional_cases()
}
\value{
A dataframe of daily French regional case counts.
}
\description{
Data from Santé Publique France https://github.com/opencovid19-fr/data
}
\examples{

## Code
get_france_regional_cases()
}
\author{
Paul Campbell @paulcampbell91
}
