% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france.R
\name{get_france_regional_cases_only_level_1}
\alias{get_france_regional_cases_only_level_1}
\title{French Regional Daily COVID-19 Count Data - Region}
\usage{
get_france_regional_cases_only_level_1()
}
\value{
A data frame of COVID cases by Region in France, ready to be used by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for France, stratified by region.
Data available at \url{https://www.data.gouv.fr/fr/datasets/donnees-relatives-aux-resultats-des-tests-virologiques-covid-19/}.
It is loaded and then sanitised.
}
