% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france.R
\name{get_france_regional_cases_with_level_2}
\alias{get_france_regional_cases_with_level_2}
\title{French Regional Daily COVID-19 Count Data - Departement}
\usage{
get_france_regional_cases_with_level_2()
}
\value{
A data.frame of COVID cases by departement in France, ready to be used by get_regional_data().
}
\description{
Extracts daily COVID-19 data for France, stratified by departement.
Data available at \url{https://www.data.gouv.fr/fr/datasets/donnees-relatives-aux-resultats-des-tests-virologiques-covid-19/}.
It is loaded and then sanitised.
}
