% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_germany_level_2_codes}
\alias{get_germany_level_2_codes}
\title{German level 2 codes (not available currently)}
\usage{
get_germany_level_2_codes()
}
\description{
German level 2 codes (not available currently)
}
