% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_germany_regional_cases.R
\name{get_germany_regional_cases_only_level_1}
\alias{get_germany_regional_cases_only_level_1}
\title{Get German daily cases by Bundesland only}
\usage{
get_germany_regional_cases_only_level_1()
}
\value{
A data.frame of COVID cases by region in Germany, ready to be used by get_regional_covid_data()
}
\description{
Fetches COVID case counts stratified by Bundesland in Germany.
This data is sourced from the Robert Koch Institute:
https://npgeo-corona-npgeo-de.hub.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0
selects and sanitises the relevant columns
}
