% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_covidregionaldata.R
\name{get_info_covidregionaldata}
\alias{get_info_covidregionaldata}
\title{Get meta information about the covidregionaldata datasets.
Use this also as a guide for setting up new country datasets}
\usage{
get_info_covidregionaldata()
}
\value{
A tibble of metadata about covidregionaldata
}
\description{
Find out function names, data returned, subregion level and raw data sources, alongside package info
}
