% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R-wrappers.R
\name{get_info_ncovutils}
\alias{get_info_ncovutils}
\title{Get meta information about the NCovUtils refresh_* data getters.}
\usage{
get_info_ncovutils()
}
\value{
A tibble of metadata about NCoVUtils
}
\description{
Find out function names, data returned, subregion level and raw data sources, alongside package info such as licence details.
}
