% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interventions_data.R
\name{get_interventions_data}
\alias{get_interventions_data}
\title{Import ACAPS Government Interventions dataset}
\usage{
get_interventions_data()
}
\value{
a dataframe of government interventions compiled by ACAPS
}
\description{
Data available here: https://data.humdata.org/dataset/acaps-covid19-government-measures-dataset#
}
\examples{

## Code
get_interventions_data
}
\author{
Paul Campbell @paulcampbell91
}
