% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso_codes.R
\name{get_iso_codes}
\alias{get_iso_codes}
\title{Get a table of ISO codes for a specified country}
\usage{
get_iso_codes(country)
}
\arguments{
\item{country}{a string with a country specified}
}
\value{
a tibble of regions and their corresponding ISO codes
}
\description{
Get a table of ISO codes for a specified country
}
