% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_italy_regional_cases.R
\name{get_italy_regional_cases}
\alias{get_italy_regional_cases}
\title{Italian Regional Case Counts}
\usage{
get_italy_regional_cases()
}
\value{
A dataframe of Italian regional case counts.
}
\description{
Italian Regional Case Counts
}
\examples{

## Code
get_italy_regional_cases
}
