% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japan.R
\name{get_japan_regional_cases}
\alias{get_japan_regional_cases}
\title{Japan regional cases, daily}
\usage{
get_japan_regional_cases()
}
\description{
Extracts regional case counts for Japan.
\href{https://mhlw-gis.maps.arcgis.com/apps/opsdashboard/index.html#/0c5d0502bbb54f9a8dddebca003631b8}{Source}.
}
