% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu.R
\name{get_jhu_data}
\alias{get_jhu_data}
\title{Download and clean data from John Hopkins Univetsity (JHU)}
\usage{
get_jhu_data()
}
\value{
A data frame of COVID cases for all countries listed by JHU,
by country and region if available
}
\description{
Downloads daily covid data from JHU
(https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data)
and formats the data
}
