% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korea.R
\name{get_korea_regional_cases}
\alias{get_korea_regional_cases}
\title{South Korea regional cases}
\usage{
get_korea_regional_cases()
}
\description{
Extract regional case counts for South Korea.
\href{http://ncov.mohw.go.kr/en}{Source}.
}
\examples{

\dontrun{
# Mapping
data <- get_korea_regional_cases()

regions <- rnaturalearth::ne_states("South Korea", returnclass = "sf")

regions_with_data <- regions \%>\%
mutate(name_de = str_replace_all(name_de, "Jeju-do", "Jeju")) \%>\%
left_join(data, by = c("name_de" = "region"))


ggplot(regions_with_data) +
geom_sf(aes(fill = cases))
}

get_korea_regional_cases
}
