% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_level_2_region_codes}
\alias{get_level_2_region_codes}
\title{Get a table of level 2 region codes (FIPS, ONS, region) for a specified country}
\usage{
get_level_2_region_codes(country)
}
\arguments{
\item{country}{a string with a country specified}
}
\value{
a tibble of regions and their corresponding level 2 region codes
}
\description{
Get a table of level 2 region codes (FIPS, ONS, region) for a specified country
}
