% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_lithuania_region_codes}
\alias{get_lithuania_region_codes}
\title{Lithuanian region codes}
\usage{
get_lithuania_region_codes()
}
\value{
(NULL) These are already presented by the
\code{get_lithuania_regional_cases_with_level_2} function
}
\description{
Lithuanian region codes
}
