% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lithuania.R
\name{get_lithuania_regional_cases_only_level_1}
\alias{get_lithuania_regional_cases_only_level_1}
\title{Lithuania Regional Daily COVID-19 Count Data - County}
\usage{
get_lithuania_regional_cases_only_level_1()
}
\value{
A data frame of COVID cases by county in Lithuania, ready to be used by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for Lithuania, stratified by county (apskritis)
and municipality (savivaldybe). Some Lithuanian municipalities share names, there being both a
Vilnius city municipality (m. sav.) and a Vilnius regional municipality (r. sav.)
Data available at \url{https://opendata.arcgis.com/datasets/45b76303953d40e2996a3da255bf8fe8_0}.
It is loaded and then sanitised.
}
