% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico.R
\name{get_mexico_inegi_codes}
\alias{get_mexico_inegi_codes}
\title{Mexico level 1 codes
Needed to join across level 1 ISO codes, full region 1 names, and INEGI state codes used in level 2 codes.}
\usage{
get_mexico_inegi_codes()
}
\description{
Mexico level 1 codes
Needed to join across level 1 ISO codes, full region 1 names, and INEGI state codes used in level 2 codes.
}
