% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_mexico_level_2_codes}
\alias{get_mexico_level_2_codes}
\title{Mexico level 2 codes (included in original function)}
\usage{
get_mexico_level_2_codes()
}
\description{
Mexico level 2 codes (included in original function)
}
