% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico.R
\name{get_mexico_regional_cases_only_level_1}
\alias{get_mexico_regional_cases_only_level_1}
\title{Mexican Regional Daily COVID-19 Count Data - Estado}
\usage{
get_mexico_regional_cases_only_level_1()
}
\value{
A data frame of COVID cases by Estado in Mexico, ready to be used by \code{\link[=get_regional_data]{get_regional_data()}}.
}
\description{
Extracts daily COVID-19 data for Mexico, stratified by Estados.
Data available at \url{https://datos.covid-19.conacyt.mx/#DownZCSV}.
It is loaded and then sanitised.
}
