% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico.R
\name{get_mexico_regional_cases_with_level_2}
\alias{get_mexico_regional_cases_with_level_2}
\title{Mexican Regional Daily COVID-19 Count Data - Municipio}
\usage{
get_mexico_regional_cases_with_level_2()
}
\value{
A data.frame of COVID cases by Municipio in Mexico, ready to be used by \code{\link[=get_regional_data]{get_regional_data()}}.
}
\description{
Extracts daily COVID-19 data for Germany, stratified by Municipio
Data available at \url{https://datos.covid-19.conacyt.mx/#DownZCSV}
It is loaded and then sanitised.
}
