% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_national_data.R
\name{get_national_data}
\alias{get_national_data}
\title{Get national-level data for countries globally, sourced from the ECDC or WHO.}
\usage{
get_national_data(country, source = "who", steps = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{country}{A character string specifying the country to get data from.
Not case dependent. Name should be the English name. For a list of
options use \code{get_available_datasets}.}

\item{source}{A character string specifying the data source: "WHO", or
"ECDC". Not case dependent. Defaults to WHO.}

\item{steps}{Logical, defaults to FALSE. Should all processing and cleaning
steps be kept and output in a list.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{additional arguments to pass to Country classes.}
}
\value{
A tibble with data related to cases, deaths, hospitalisations,
recoveries and testing.
}
\description{
Gets raw data using the source-specific function. Includes ISO
country codes. Then adds columns which were missing from the raw data
(calculating cumulative counts from new dailies and vice versa), cleans
and sanitises further. Adds rows and columns of NA values so that data is
in a standard format.
}
\examples{
\dontrun{
get_national_data("Italy", source = "who")
}
}
