% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UK.R, R/shared-methods.R
\name{Uk}
\alias{Uk}
\alias{DataClass}
\title{United Kingdom Class for downloading, cleaning and processing notification
data.}
\source{
https://coronavirus.data.gov.uk/details/download #nolint
}
\description{
Country specific information for downloading, cleaning
and processing covid-19 region data for an example Country.

Acts as parent class for individual country objects,
allowing them to access general methods.
}
\details{
Inherits from \code{DataClass}

All countries have shared methods for extracting region codes,
downloading, processing, and returning data.
}
\examples{
\dontrun{
region <- Italy$new(verbose = TRUE, steps = TRUE)
region$download()
region$clean()
region$process()
region$return()
}
}
\section{Super class}{
\code{\link[covidregionaldata:DataClass]{covidregionaldata::DataClass}} -> \code{Uk}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level_1_region}}{the level 1 region name.}

\item{\code{level_2_region}}{the level 2 region name}

\item{\code{data_url}}{link to raw data}

\item{\code{source_data_cols}}{existing columns within the raw data}

\item{\code{query_filters}}{Set what filters to use to query the data}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{query_filters}}{Set what filters to use to query the data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_region_codes}{\code{Uk$get_region_codes()}}
\item \href{#method-download}{\code{Uk$download()}}
\item \href{#method-clean}{\code{Uk$clean()}}
\item \href{#method-clean_level_1}{\code{Uk$clean_level_1()}}
\item \href{#method-new}{\code{Uk$new()}}
\item \href{#method-clone}{\code{Uk$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="DataClass" data-id="process">}\href{../../covidregionaldata/html/DataClass.html#method-process}{\code{covidregionaldata::DataClass$process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="DataClass" data-id="return">}\href{../../covidregionaldata/html/DataClass.html#method-return}{\code{covidregionaldata::DataClass$return()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_region_codes"></a>}}
\if{latex}{\out{\hypertarget{method-get_region_codes}{}}}
\subsection{Method \code{get_region_codes()}}{
Specific function for getting region codes for UK .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$get_region_codes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$download()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\if{latex}{\out{\hypertarget{method-clean}{}}}
\subsection{Method \code{clean()}}{
UK specific cleaning, directs to level 1 or level 2
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$clean(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{pass additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_level_1"></a>}}
\if{latex}{\out{\hypertarget{method-clean_level_1}{}}}
\subsection{Method \code{clean_level_1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$clean_level_1(nhsregions = FALSE, release_date = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the country
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The args passed by \link{general_init}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{country}}{name of country to fetch data for}

\item{\code{region}}{data frame for requested region}

\item{\code{data_url}}{link to raw data}

\item{\code{level}}{target region level}

\item{\code{totals}}{Boolean. If TRUE, returns totalled data per region
up to today's date.}

\item{\code{localise}}{Boolean. Should region names be localised.}

\item{\code{verbose}}{Boolean. Display information at various stages.}

\item{\code{steps}}{Boolean. Keep data from each processing step.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_region_codes}{\code{DataClass$get_region_codes()}}
\item \href{#method-download}{\code{DataClass$download()}}
\item \href{#method-process}{\code{DataClass$process()}}
\item \href{#method-return}{\code{DataClass$return()}}
\item \href{#method-clone}{\code{DataClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_region_codes"></a>}}
\if{latex}{\out{\hypertarget{method-get_region_codes}{}}}
\subsection{Method \code{get_region_codes()}}{
General function for getting region codes for given region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$get_region_codes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
General function for downloading raw data.
Shared regional dataset processing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$download()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process"></a>}}
\if{latex}{\out{\hypertarget{method-process}{}}}
\subsection{Method \code{process()}}{
General function to processes regional data.
Dynamically works for level 1 and level 2 regions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$process()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return"></a>}}
\if{latex}{\out{\hypertarget{method-return}{}}}
\subsection{Method \code{return()}}{
Optional region specific return changes.
Designed to be called after \code{process}. For most datasets a
custom method should not be needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$return()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
