% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/russia.R
\name{get_russia_regional_cases}
\alias{get_russia_regional_cases}
\title{Get Russia daily cases}
\usage{
get_russia_regional_cases()
}
\value{
A dataframe of case counts in Russian regions
}
\description{
Fetches COVID case counts by region
This data is sourced from https://github.com/grwlf/COVID-19_plus_Russia#data-sources
}
\examples{


\dontrun{


regions <- rnaturalearth::ne_states(geounit = "Russia", returnclass = "sf")
data <- get_russia_regional_cases() \%>\%
  dplyr::filter(date == max(date))
regions_with_data <- dplyr::left_join(regions, data, by = "iso_3166_2")
regions_with_data \%>\%
  ggplot2::ggplot(ggplot2::aes(fill = cases)) +
  ggplot2::geom_sf() +
  ggplot2::coord_sf(xlim = c(20, max(sf::st_coordinates(regions_with_data))))
  }
}
