% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_southafrica_region_codes}
\alias{get_southafrica_region_codes}
\title{South Africa region codes (NULL - they're in the raw data already)}
\usage{
get_southafrica_region_codes()
}
\description{
South Africa region codes (NULL - they're in the raw data already)
}
