% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southafrica.R
\name{get_southafrica_regional_cases_only_level_1}
\alias{get_southafrica_regional_cases_only_level_1}
\title{South Africa Regional Daily COVID-19 Count Data}
\usage{
get_southafrica_regional_cases_only_level_1()
}
\value{
A data frame of COVID cases by province in South Africa ready to be used by \code{get_regional_data()}.
}
\description{
Fetches daily COVID-19 data
Data is available at \url{https://github.com/dsfsi/covid19za}.
}
