% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spain_regional_cases.R
\name{get_spain_regional_cases}
\alias{get_spain_regional_cases}
\title{Spain regional cases, daily
Source: https://www.mscbs.gob.es/profesionales/saludPublica/ccayes/alertasActual/nCov-China/situacionActual.htm
NB. PDF table is not produced consistently each day, so code (lines 18 & 20) may need adjusting
Alternative source here: https://covid19.isciii.es/}
\usage{
get_spain_regional_cases()
}
\value{
A dataframe of daily Spanish regional case counts.
}
\description{
Spain regional cases, daily
Source: https://www.mscbs.gob.es/profesionales/saludPublica/ccayes/alertasActual/nCov-China/situacionActual.htm
NB. PDF table is not produced consistently each day, so code (lines 18 & 20) may need adjusting
Alternative source here: https://covid19.isciii.es/
}
\examples{
Code
}
