% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_cases.R
\name{get_total_cases}
\alias{get_total_cases}
\title{Get Total Cases by Country}
\usage{
get_total_cases()
}
\value{
Returns a data frame of variables along with their most recent value.
}
\description{
Get Total Cases by Country
}
\examples{


get_total_cases()
}
