% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{get_uk_data}
\alias{get_uk_data}
\title{Get UK data - helper function to get data for a single valid area type}
\usage{
get_uk_data(filters, release_date = NULL)
}
\arguments{
\item{filters}{Query filters for UK data e.g. "areaType=nation"}

\item{release_date}{Date data was released. Default is to extract latest release.
Dates should be in the format "yyyy-mm-dd".}
}
\value{
A dataframe with all variables available in public UK data
}
\description{
UK data download helper. Data extracted using:
https://coronavirus.data.gov.uk/details/download
}
