% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso_codes.R
\name{get_uk_iso_codes}
\alias{get_uk_iso_codes}
\title{UK ISO codes (NULL - they're in the raw data already)}
\usage{
get_uk_iso_codes()
}
\description{
UK ISO codes (NULL - they're in the raw data already)
}
