% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_uk_level_2_codes}
\alias{get_uk_level_2_codes}
\title{UK level 2 codes (ONS) (Included in original function)}
\usage{
get_uk_level_2_codes()
}
\description{
UK level 2 codes (ONS) (Included in original function)
}
