% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uk_nhs_region_cases.R
\name{get_uk_nhs_region_cases}
\alias{get_uk_nhs_region_cases}
\title{Get English daily cases by NHS region}
\usage{
get_uk_nhs_region_cases()
}
\value{
A dataframe of case counts in English NHS regions
}
\description{
Get English daily cases by NHS region
}
\examples{
get_uk_nhs_region_cases
}
