% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_uk_region_codes}
\alias{get_uk_region_codes}
\title{UK region codes (NULL - they're in the raw data already)}
\usage{
get_uk_region_codes()
}
\description{
UK region codes (NULL - they're in the raw data already)
}
