% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{get_uk_regional_cases_with_level_2}
\alias{get_uk_regional_cases_with_level_2}
\title{Get UK daily cases}
\usage{
get_uk_regional_cases_with_level_2()
}
\arguments{
\item{geography}{Character string identifying which part of the UK to extract and at what scale
Defaults to "all countries". Options are: "all countries", "all regions", "England", "Scotland", "Northern Ireland", "Wales".
Note that England always returns 9 English regions, and Nothern Ireland always returns as one country}
}
\value{
A dataframe of case counts. By default returns cases in English regions, and Scotland, Wales, and Northern Ireland
}
\description{
Get UK cases by country or region
}
