% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_us_level_2_codes}
\alias{get_us_level_2_codes}
\title{US level 2 codes (FIPS) (Included in original function)}
\usage{
get_us_level_2_codes()
}
\description{
US level 2 codes (FIPS) (Included in original function)
}
