% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_us_regional_cases.R
\name{get_us_regional_cases}
\alias{get_us_regional_cases}
\title{Fetch daily COVID cases by state for United States of America}
\usage{
get_us_regional_cases()
}
\value{
A data.frame of COVID cases by region in the US
}
\description{
Fetches daily COVID cases by state collated by the
John Hopkins Centre for Systems Science and Engineering (JHU CSSE)
Data is available at https://github.com/CSSEGISandData/COVID-19
}
\examples{
get_us_regional_cases

\dontrun{

 regions <- rnaturalearth::ne_states("United States of America", returnclass = "sf")

 cases <- get_us_regional_cases()

 regions \%>\% dplyr::filter(name \%in\% cases$region) \%>\%
 ggplot2::ggplot() + ggplot2::geom_sf()

}
}
