% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa.R
\name{get_us_regional_cases_only_level_1}
\alias{get_us_regional_cases_only_level_1}
\title{Fetch daily COVID cases by state for United States of America}
\usage{
get_us_regional_cases_only_level_1()
}
\value{
A data.frame of COVID cases by region in the US
}
\description{
Fetches daily COVID cases county collated by the
New York Times
Data is available at https://github.com/nytimes/covid-19-data
}
