% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_who_cases.R
\name{get_who_cases}
\alias{get_who_cases}
\title{Download the most recent WHO case data}
\usage{
get_who_cases(country = NULL, daily = FALSE, cache = NULL)
}
\arguments{
\item{country}{Character string of current name}

\item{daily}{Logical, defaults to \code{FALSE}. Should cases}

\item{cache}{A \code{memoise} cache.
counts be daily (\code{TRUE}) or cumulative (\code{FALSE}).}
}
\value{
A datatable of either all WHO data by date or cases in a specific country by date
}
\description{
Downloads the latest WHO case data.It uses \code{memoise} to cache the
results locally. To clear the cache and redownload the data use \code{reset_cache}. The cache will be located
in the directory at which the function is run.
}
\examples{

get_who_cases
}
