% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_covid_data.R
\name{get_wide_format_regional_covid_data}
\alias{get_wide_format_regional_covid_data}
\title{Return regional, daily-updated data for a given country in wide (time-series) format.}
\usage{
get_wide_format_regional_covid_data(country, include_level_2_regions = FALSE)
}
\arguments{
\item{country}{Character String specifying the country to get data from. For options see the documentation.}
}
\value{
A data.frame with data related to cases, deaths, hospitalisations, recoveries and testing for regions within the given country.
}
\description{
Gets COVID-19 data related to cases, deaths, hospitalisations, recoveries and testing for sub-regions of a given country in wide format.
}
\examples{

\dontrun{

 get_wide_format_regional_covid_data(country = "canada")

}
}
