% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glue_level}
\alias{glue_level}
\title{Glue the spatial level into a variable name}
\usage{
glue_level(level)
}
\arguments{
\item{level}{A character string indicating the current level.}
}
\value{
A string in the form "level_1_region".
}
\description{
Glue the spatial level into a variable name
}
\concept{utility}
