% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{italy_codes}
\alias{italy_codes}
\title{Region Codes for Italy Dataset.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
italy_codes
}
\value{
A nested tibble of region codes and related information.
}
\description{
Details of the region codes used for the Italy dataset.
}
\keyword{datasets}
