% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{JHU_codes}
\alias{JHU_codes}
\title{Region Codes for JHU Dataset. Taken from the region codes provided as
part of the WHO dataset.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4193 rows and 2 columns.
}
\usage{
JHU_codes
}
\value{
A tibble of region codes and related information.
}
\description{
The region codes for JHU
}
\keyword{datasets}
