% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json_reader}
\alias{json_reader}
\title{Custom JSON reading function}
\usage{
json_reader(file, verbose = FALSE, ...)
}
\arguments{
\item{file}{A URL or filepath to a JSON}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{extra parameters to be passed to jsonlite::fromJSON}
}
\value{
A data table
}
\description{
Checks for use of memoise and then uses vroom::vroom.
}
\concept{utility}
