% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left_join_region_codes}
\alias{left_join_region_codes}
\title{Custom left_join function}
\usage{
left_join_region_codes(data, region_codes_table, by = NULL, ...)
}
\arguments{
\item{data}{a data table}

\item{region_codes_table}{a table of region codes which will be left_join (optionally NULL)}

\item{by}{see dplyr::left_join() description of by parameter}

\item{...}{optional arguments passed into dplyr::left_join()}
}
\value{
A data table
}
\description{
Checks if table that is being added is NULL and then uses left_join
}
