% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lithuania_codes}
\alias{lithuania_codes}
\title{Region Codes for Lithuania Dataset.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 61 rows and 6 columns.
}
\usage{
lithuania_codes
}
\value{
A tibble of region codes and related information,
including ISO 3166:2 codes for counties (apskritis)
and municipalities (savivaldybe), and noting which
municipalities are city municipalities or regional
municipalities.
}
\description{
The region codes for Lithuania
}
\keyword{datasets}
