% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_verbose}
\alias{message_verbose}
\title{Wrapper for message}
\usage{
message_verbose(verbose = TRUE, ...)
}
\arguments{
\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments passed to \code{message}.}
}
\description{
A wrapper for \code{message} that only prints output when
\code{verbose = TRUE}.
}
\concept{utility}
