% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico.R
\name{mexico_api}
\alias{mexico_api}
\title{Download Mexico data across levels}
\usage{
mexico_api(level = "estados", verbose = TRUE)
}
\arguments{
\item{level}{A character string indicating the target administrative level
to download. Supports "estados" and "municipio".}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}
}
\description{
Data download function for Mexico data. To get the latest data
use a PHP script from the website.
}
\author{
Sam Abbott
}
