% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-methods.R
\name{new_covidregionaldata}
\alias{new_covidregionaldata}
\title{Construct a covidregionaldata class object}
\usage{
new_covidregionaldata(country = character(), level = "1", verbose = TRUE)
}
\arguments{
\item{country}{A character string specifying the country to get data from.
Not case dependent. Name should be the English name. For a list of
options see \code{available_datasets}.}

\item{level}{A character string indicating the target administrative level
of the data with the default being "1". Currently supported options are
level 1 ("1) and level 2 ("2"). See \code{available_datasets} for supported
options by dataset.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}
}
\value{
A list containing target country name, administrative level of
the data, and a list of information about the region codes. The list
well be assigned a class start "crd_" and containing the target country
and administrative level to control method dispatch.
}
\description{
Constructs a \code{covidregionaldata} class object for the
target dataset of interest (see return). Returns an error if the requested
dataset is not supported.
}
\examples{
# initialise data downloading in the UK
new_covidregionaldata("uk", "1")
# initialise data downloading for mexico
new_covidregionaldata("mexico", "1")
# initialise data download from the ECDC source
new_covidregionaldata("ecdc")
}
\author{
Sam Abbott
}
