% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-methods.R
\name{process_regional.crd_level_1}
\alias{process_regional.crd_level_1}
\title{Shared Regional Dataset Processing for Level 1 Data}
\usage{
\method{process_regional}{crd_level_1}(region, totals = FALSE, localise = FALSE, verbose = TRUE)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{totals}{Logical, defaults to FALSE. If TRUE, returns totalled
data per region up to today's date. If FALSE, returns the full dataset
stratified by date and region.}

\item{localise}{Logical, defaults to TRUE. Should region names be localised.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}
}
\description{
Shared regional data cleaning designed to be called
after \code{clean_regional} for level 1 regions.
}
\examples{
\dontrun{
mexico <- new_covidregionaldata("mexico")
mexico <- download_regional(mexico)
mexico <- clean_regional(mexico)
process_regional(mexico)$processed
}
}
\author{
Sam Abbott
}
