% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{process_regional}
\alias{process_regional}
\title{Shared regional dataset processing}
\usage{
process_regional(region, verbose = TRUE, ...)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments to pass to dataset specific methods}
}
\description{
Shared regional data cleaning designed to be called
after \code{clean_regional}.
}
\author{
Sam Abbott
}
