% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{process_regional_internal}
\alias{process_regional_internal}
\title{Internal Shared Regional Dataset Processing}
\usage{
process_regional_internal(
  region,
  group_vars,
  totals = FALSE,
  localise = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{region}{A given Country class object to process, e.g. \code{Italy()}}

\item{group_vars}{A character vector of grouping variables. It is assumed
that the first entry indicates the main region variable and the second
indicates the code for this variable.}

\item{totals}{Logical, defaults to \code{FALSE}. If `TRUE``, returns totalled
data per region up to today's date. If FALSE, returns the full dataset
stratified by date and region.}

\item{localise}{Logical, defaults to \code{TRUE}. Should region names be
localised.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}
}
\description{
Internal shared regional data cleaning designed to be called
by \code{process}.
}
