% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_france_regional_cases.R
\name{read_spf_regions_yaml}
\alias{read_spf_regions_yaml}
\title{Read SPF yaml file to dataframe}
\usage{
read_spf_regions_yaml(path)
}
\arguments{
\item{path}{path to daily SPF covid-19 yaml file}
}
\description{
Read SPF yaml file to dataframe
}
