% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_afghanistan}
\alias{refresh_covidregionaldata_afghanistan}
\title{Get daily Afghan COVID-19 count data by Province (Wilayat)}
\usage{
refresh_covidregionaldata_afghanistan()
}
\value{
A tibble of COVID cases by province in Afghanistan.
}
\description{
Fetches  COVID-19 count data, stratified by date and province.
Data sourced from https://docs.google.com/spreadsheets/d/1F-AMEDtqK78EA6LYME2oOsWQsgJi4CT3V_G4Uo-47Rg/export?format=csv.
}
