% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_belgium}
\alias{refresh_covidregionaldata_belgium}
\title{Get daily Belgian COVID-19 count data by Region}
\usage{
refresh_covidregionaldata_belgium()
}
\value{
A tibble of COVID cases by province in Belgium.
}
\description{
Fetches COVID-19 count data, stratified by date and region
Data sourced from https://epistat.sciensano.be/Data (3 different datasets - MORT, AGE_SEX and HOSP used)
}
