% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_brazil}
\alias{refresh_covidregionaldata_brazil}
\title{Get daily Brazilian COVID-19 count data by State (Estado)}
\usage{
refresh_covidregionaldata_brazil()
}
\value{
A tibble of COVID cases by province in Brazil
}
\description{
Fetches COVID-19 count data, stratified by date and state
Data sourced from https://raw.githubusercontent.com/wcota/covid19br/master/cases-brazil-cities-time.csv.
}
