% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_canada}
\alias{refresh_covidregionaldata_canada}
\title{Get daily Canada COVID-19 count data by Province/Territory}
\usage{
refresh_covidregionaldata_canada()
}
\value{
A tibble of COVID cases by province in Canada.
}
\description{
Fetches COVID-19 count data, stratified by date and province.
Data sourced from https://health-infobase.canada.ca/src/data/covidLive/covid19.csv.
}
