% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_colombia}
\alias{refresh_covidregionaldata_colombia}
\title{Get daily Colombian COVID-19 count data by Department (Departamento).}
\usage{
refresh_covidregionaldata_colombia()
}
\value{
A tibble of COVID cases by province in Colombia.
}
\description{
Fetches  COVID-19 count data, stratified by date and region.
Data sourced from https://raw.githubusercontent.com/ideascol/covid19/master/data/data_dptos_trend.csv.
}
