% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_germany}
\alias{refresh_covidregionaldata_germany}
\title{Get daily German COVID-19 count data by State (Bundesland)}
\usage{
refresh_covidregionaldata_germany()
}
\value{
A tibble of COVID cases by province in Germany.
}
\description{
Fetches  COVID-19 count data, stratified by date and state.
Data sourced from https://opendata.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0.csv.
}
