% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_india}
\alias{refresh_covidregionaldata_india}
\title{Get daily Indian COVID-19 count data by State/Unified Territory}
\usage{
refresh_covidregionaldata_india()
}
\value{
A tibble of COVID cases by province in India
}
\description{
Fetches COVID-19 count data, stratified by date and state.
Data sourced from https://api.covid19india.org/csv/latest/state_wise_daily.csv.
}
