% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_italy}
\alias{refresh_covidregionaldata_italy}
\title{Get daily Italian COVID-19 count data by Region (Regioni).}
\usage{
refresh_covidregionaldata_italy()
}
\value{
A tibble of COVID cases by province in Italy.
}
\description{
Fetches  COVID-19 count data, stratified by date and region.
Data sourced from https://raw.githubusercontent.com/pcm-dpc/COVID-19/master/dati-regioni/dpc-covid19-ita-regioni-\emph{\strong{date}}.csv.
}
