% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_russia}
\alias{refresh_covidregionaldata_russia}
\title{Get daily Russian COVID-19 count data by Russian region.}
\usage{
refresh_covidregionaldata_russia()
}
\value{
A tibble of COVID cases by province in Russia.
}
\description{
Fetches  COVID-19 count data, stratified by date and region.
Data sourced from https://raw.githubusercontent.com/grwlf/COVID-19_plus_Russia/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_RU.csv.
}
