% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_uk}
\alias{refresh_covidregionaldata_uk}
\title{Get daily UK COVID-19 count data by EU-defined region}
\usage{
refresh_covidregionaldata_uk()
}
\value{
A tibble of COVID cases by EU region in UK.
}
\description{
Fetches  COVID-19 count data, stratified by date and region.
Data sourced from https://coronavirus.data.gov.uk/downloads/csv/coronavirus-cases_latest.csv and
https://raw.githubusercontent.com/tomwhite/covid-19-uk-data/master/data/covid-19-cases-uk.csv.
}
