% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R_wrappers.R
\name{refresh_covidregionaldata_usa}
\alias{refresh_covidregionaldata_usa}
\title{Get daily USA COVID-19 count data by state.}
\usage{
refresh_covidregionaldata_usa()
}
\value{
A tibble of COVID cases by state in USA
}
\description{
Fetches  COVID-19 count data, stratified by date and region.
Data sourced from https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-states.csv.
}
