% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19R-wrappers.R
\name{refresh_ncovutils_afghanistan}
\alias{refresh_ncovutils_afghanistan}
\title{Get daily Afghan COVID-19 count data by Province (Wilåyat)}
\usage{
refresh_ncovutils_afghanistan()
}
\value{
A tibble of COVID cases by province in Afghanistan.
}
\description{
Fetches  COVID-19 count data, stratified by date and province.
Data sourced from https://docs.google.com/spreadsheets/d/1F-AMEDtqK78EA6LYME2oOsWQsgJi4CT3V_G4Uo-47Rg/export?format=csv.
}
