% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{region_codes}
\alias{region_codes}
\title{Region Codes for Each Dataset.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
region_codes
}
\value{
A nested tibble of region codes and related information.
}
\description{
Details of the region codes used by each dataset.
}
\keyword{datasets}
