% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{region_dispatch}
\alias{region_dispatch}
\title{Control Grouping Variables used in process_internal}
\usage{
region_dispatch(level, all_levels, region_names, region_codes)
}
\arguments{
\item{level}{A character string indicating the current level.}

\item{all_levels}{A character vector indicating all the levels supported.}

\item{region_names}{A named list of region names named after the levels
supported.}

\item{region_codes}{A named list of region codes named after the levels
supported.}
}
\description{
Controls the grouping variables used in
\code{process_internal} based on the supported regions present in the
class.
}
\concept{utility}
