% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_region_column}
\alias{rename_region_column}
\title{Helper to rename the region column in each dataset to the correct name for each country.}
\usage{
rename_region_column(data, country)
}
\arguments{
\item{data}{a data frame with a region_level_1 column and optionally a region_level_2 column}

\item{country}{a string with the country of interest}
}
\value{
a tibble with the column renamed to a sensible name
}
\description{
The package relies on column name 'region' during processing but this often isn't the most sensible name for the column
(e.g. state makes more sense for USA). This simply renames the column as the final step in processing before returning data to the user.
}
