% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_datasets.R
\name{render_available_datasets}
\alias{render_available_datasets}
\title{Make table of supported data sets and store as package data}
\usage{
render_available_datasets()
}
\description{
The function searches the environment for R6 class objects and
extracts the summary information from the available classes using
their \code{summary} methods. In practice this means that it can be used
to indicate supported data sets.
}
\examples{
# see all available datasets
render_available_datasets()
}
\seealso{
Data interface functions
\code{\link{CountryDataClass}},
\code{\link{DataClass}},
\code{\link{get_available_datasets}()},
\code{\link{get_national_data}()},
\code{\link{get_regional_data}()},
\code{\link{initialise_dataclass}()}
}
\concept{interface}
