% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_cache.R
\name{reset_cache}
\alias{reset_cache}
\title{Reset Cache and Update all Local Data}
\usage{
reset_cache(refresh_data = FALSE)
}
\arguments{
\item{refresh_data}{Logical defaults to \code{FALSE}. Should all data sources be refreshed once the cache has been
removed.}
}
\value{
Null
}
\description{
Reset Cache and Update all Local Data
}
\examples{

## Code
reset_cache
}
