% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdc.R
\name{return_regional.crd_ecdc_1}
\alias{return_regional.crd_ecdc_1}
\title{WHO Return Changes}
\usage{
\method{return_regional}{crd_ecdc_1}(region, steps = FALSE)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{steps}{Logical, defaults to FALSE. Should all processing and cleaning
steps be kept and output in a list.}
}
\description{
Specific return settings for the WHO dataset.
}
\examples{
\dontrun{
ecdc <- new_covidregionaldata("ecdc")
ecdc <- download_regional(ecdc)
ecdc <- clean_regional(ecdc)
ecdc <- process_regional(ecdc)
return_regional(ecdc)
}
}
\author{
Sam Abbott
}
