% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{return_regional.crd_who_1}
\alias{return_regional.crd_who_1}
\title{WHO Return Changes}
\usage{
\method{return_regional}{crd_who_1}(region, steps = FALSE)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{steps}{Logical, defaults to FALSE. Should all processing and cleaning
steps be kept and output in a list.}
}
\description{
Specific return settings for the WHO dataset.
}
\examples{
\dontrun{
who <- new_covidregionaldata("who")
who <- download_regional(who)
who <- clean_regional(who)
who <- process_regional(who)
return_regional(who)
}
}
\author{
Sam Abbott
}
