% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-methods.R
\name{return_regional.default}
\alias{return_regional.default}
\title{Default Region Specific Return Changes}
\usage{
\method{return_regional}{default}(region, steps = FALSE)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{steps}{Logical, defaults to FALSE. Should all processing and cleaning
steps be kept and output in a list.}
}
\description{
Optional region specific return changes. Designed to be called
after \code{process_regional}. For most datasets a custom method should not be
needed.
}
\author{
Sam Abbott
}
