% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-methods.R
\name{return_regional}
\alias{return_regional}
\title{Region Specific Return Settings}
\usage{
return_regional(region, steps = FALSE)
}
\arguments{
\item{region}{A list as produced by \code{new_covidregionaldata}.}

\item{steps}{Logical, defaults to FALSE. Should all processing and cleaning
steps be kept and output in a list.}
}
\description{
Region specific return settings. Designed to be called
after \code{process_regional}. For most datasets a custom method should not be
needed.
}
\author{
Sam Abbott
}
