% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgamma_with_mean_sd.R
\name{rgamma_with_mean_sd}
\alias{rgamma_with_mean_sd}
\title{Mean and standard deviation parameterised gamma sampling function}
\usage{
rgamma_with_mean_sd(n, mean, sd)
}
\arguments{
\item{n}{Numeric, the number of samples to take.}

\item{mean}{Numeric, the mean of the distribution.}

\item{sd}{Numeric the standard deviation of the distribution.}
}
\value{
A vector of samples from the gamma distribution
}
\description{
Mean and standard deviation parameterised gamma sampling function
}
\examples{

## Example 
sample <- rgamma_with_mean_sd(1000, 2, 3)


mean(sample) 

sd(sample)

hist(sample)

## Code 
rgamma_with_mean_sd
}
