% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_optional_processing_fns}
\alias{run_optional_processing_fns}
\title{Optional processing steps to run}
\usage{
run_optional_processing_fns(data, process_fns)
}
\arguments{
\item{data}{A data table}

\item{process_fns}{array, additional functions to be called after default
processing steps}
}
\description{
user supplied processing steps which are run after default steps
}
\seealso{
Functions used in the processing pipeline
\code{\link{process_internal}()},
\code{\link{run_default_processing_fns}()}
}
\concept{processing}
\concept{utility}
