% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_process_steps}
\alias{run_process_steps}
\title{Get totals data given the time series data.}
\usage{
run_process_steps(data, process_options)
}
\arguments{
\item{data}{A data table}

\item{process_options}{list of options for processing}
}
\description{
Get totals data given the time series data.
}
