% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rweibull_with_mean_sd.R
\name{rweibull_with_mean_sd}
\alias{rweibull_with_mean_sd}
\title{Mean and standard deviation parameterised Weibull sampling function}
\usage{
rweibull_with_mean_sd(n, mean, sd)
}
\arguments{
\item{n}{Numeric, the number of samples to take.}

\item{mean}{Numeric, the mean of the distribution.}

\item{sd}{Numeric the standard deviation of the distribution.}
}
\value{
A vector of samples from the Weibull distribution
}
\description{
Mean and standard deviation parameterised Weibull sampling function
}
\examples{

## Example 
sample <- rweibull_with_mean_sd(1000, 3, 5)

mean(sample)

sd(sample)


hist(sample)
## Code 
rweibull_with_mean_sd
}
