% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_negative_values_to_zero}
\alias{set_negative_values_to_zero}
\title{Set negative data to 0}
\usage{
set_negative_values_to_zero(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data.frame with all relevant data > 0.
}
\description{
Set data values to 0 if they are negative in a dataset. Data in the datasets should always be > 0.
}
