% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_countries.R
\name{show_countries}
\alias{show_countries}
\title{Get countries data is avaliable for}
\usage{
show_countries()
}
\value{
A tibble of avaliable countries and the region level
data is avaliable for
}
\description{
Show what countries have what level data avaliable.
The function uses the region_codes file to find countries.
}
