% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\name{stop_using_memoise}
\alias{stop_using_memoise}
\title{Stop using useMemoise}
\usage{
stop_using_memoise()
}
\description{
Sets useMemoise in options to NULL, meaning memoise isn't used when reading data in
}
