% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{test_download}
\alias{test_download}
\title{Test download method works correctly}
\usage{
test_download(self, download, data_name, snapshot_path)
}
\arguments{
\item{self}{The DataClass object (R6Class) object to perform checks on.}

\item{download}{logical check to download or use a snapshot of the data}

\item{data_name}{character_array The name of the class and level to
check}

\item{snapshot_path}{character_array the path to save the downloaded
snapshot to.}
}
\description{
Test data can be downloaded if download = TRUE, or a requested
snapshot file is not found, and store a snap shot at the snapshot_path. If an
existing snapshot file is found then just load this data to use in future
tests.
}
