% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{totalise_data}
\alias{totalise_data}
\title{Get totals data given the time series data.}
\usage{
totalise_data(data, include_level_2_regions)
}
\arguments{
\item{data}{a data table}

\item{include_level_2_regions}{Boolean. Are level 2 regions included in the data}
}
\value{
A data table, totalled up
}
\description{
Get totals data given the time series data.
}
