% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uk_authority_lookup_table}
\alias{uk_authority_lookup_table}
\title{Region Codes for UK Dataset.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 429 rows and 4 columns.
}
\usage{
uk_authority_lookup_table
}
\value{
A tibble of region codes and related information.
}
\description{
The uk authority look table for providing region codes used for
level 2 UK data.
}
\keyword{datasets}
