% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USA.R
\name{USA}
\alias{USA}
\title{USA Class for downloading, cleaning and processing notification data}
\source{
https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-states.csv # nolint

https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-counties.csv # nolint
}
\description{
Country specific information for downloading, cleaning
and processing covid-19 region data for USA.
}
\details{
Inherits from \code{DataClass}
}
\examples{
\dontrun{
region <- USA$new(verbose = TRUE, steps = TRUE)
region$download()
region$clean()
region$process()
region$return()
}
}
\section{Super class}{
\code{\link[covidregionaldata:DataClass]{covidregionaldata::DataClass}} -> \code{USA}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level_1_region}}{the level 1 region name.}

\item{\code{level_2_region}}{the level 2 region name.}

\item{\code{data_url}}{link to raw data for level 1}

\item{\code{source_data_cols}}{existing columns within the raw data}

\item{\code{data_url_level_2}}{link to raw data for level 2}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-download}{\code{USA$download()}}
\item \href{#method-clean}{\code{USA$clean()}}
\item \href{#method-clean_level_1}{\code{USA$clean_level_1()}}
\item \href{#method-clean_level_2}{\code{USA$clean_level_2()}}
\item \href{#method-new}{\code{USA$new()}}
\item \href{#method-clone}{\code{USA$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="DataClass" data-id="get_region_codes">}\href{../../covidregionaldata/html/DataClass.html#method-get_region_codes}{\code{covidregionaldata::DataClass$get_region_codes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="DataClass" data-id="process">}\href{../../covidregionaldata/html/DataClass.html#method-process}{\code{covidregionaldata::DataClass$process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="DataClass" data-id="return">}\href{../../covidregionaldata/html/DataClass.html#method-return}{\code{covidregionaldata::DataClass$return()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
USA-specific function for downloading raw data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USA$download()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\if{latex}{\out{\hypertarget{method-clean}{}}}
\subsection{Method \code{clean()}}{
USA specific state level data cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USA$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_level_1"></a>}}
\if{latex}{\out{\hypertarget{method-clean_level_1}{}}}
\subsection{Method \code{clean_level_1()}}{
USA Specific State Level Data Cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USA$clean_level_1()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_level_2"></a>}}
\if{latex}{\out{\hypertarget{method-clean_level_2}{}}}
\subsection{Method \code{clean_level_2()}}{
USA Specific County Level Data Cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USA$clean_level_2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the country
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USA$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The args passed by \link{general_init}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
