% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vietnam_codes}
\alias{vietnam_codes}
\title{Region Codes for Vietnam Dataset.}
\format{
An object of class \code{data.frame} with 63 rows and 2 columns.
}
\usage{
vietnam_codes
}
\value{
A tibble of region codes and related information.
}
\description{
The region codes for Viet Nam
}
\keyword{datasets}
