% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_extra_na_cols}
\alias{add_extra_na_cols}
\title{Add extra columns filled with NA values to a dataset.}
\usage{
add_extra_na_cols(data)
}
\arguments{
\item{data}{a data table}
}
\value{
a tibble with relevant NA columns added
}
\description{
Adds extra columns filled with NAs to a dataset. This ensures that all datasets from the covidregionaldata package return datasets
of the same underlying structure (i.e. same columns).
}
