% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{calculate_columns_from_existing_data}
\alias{calculate_columns_from_existing_data}
\title{Cumulative counts from daily counts or daily counts from cumulative,
dependent on which columns already exist}
\usage{
calculate_columns_from_existing_data(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame with extra columns if required
}
\description{
Checks which columns are missing (cumulative/daily counts) and
if one is present and the other not then calculates the second from the
first.
}
\concept{utility}
