% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{expect_clean_cols}
\alias{expect_clean_cols}
\title{Expect clean columns}
\usage{
expect_clean_cols(data, level)
}
\arguments{
\item{data}{The data to check}
}
\description{
Expect data has cleaned columns. Inherited by child
classes so tests run through each class.
}
