% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{expect_clean_cols}
\alias{expect_clean_cols}
\title{Test clean columns contain the correct data and types}
\usage{
expect_clean_cols(data, level)
}
\arguments{
\item{data}{The clean data to check}

\item{level}{character_array the level of the data to check}
}
\description{
Checks the date column is an s3 class and that region level
column is a character in the cleaned data (data$clean)
}
\seealso{
Functions used for testing data is cleaned and processed correctly
\code{\link{expect_columns_contain_data}()},
\code{\link{expect_processed_cols}()},
\code{\link{test_cleaning}()},
\code{\link{test_download_JSON}()},
\code{\link{test_download}()},
\code{\link{test_processing}()},
\code{\link{test_return}()}
}
\concept{tests}
