% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{expect_columns_contain_data}
\alias{expect_columns_contain_data}
\title{Test that cleaned columns contain data/}
\usage{
expect_columns_contain_data(DataClass_obj)
}
\arguments{
\item{DataClass_obj}{The DataClass object (R6Class) to perform checks on.
Must be a \code{DataClass} or \code{DataClass} child object.}
}
\description{
Checks that cleaned columns cases, deaths, recovered and test
(new and total) are not entirely composed of NAs.
}
\seealso{
Functions used for testing data is cleaned and processed correctly
\code{\link{expect_clean_cols}()},
\code{\link{expect_processed_cols}()},
\code{\link{test_cleaning}()},
\code{\link{test_download}()},
\code{\link{test_processing}()},
\code{\link{test_return}()}
}
\concept{tests}
