% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{expect_processed_cols}
\alias{expect_processed_cols}
\title{Test that processed columns contain the correct data and types}
\usage{
expect_processed_cols(data, level = "1", localised = TRUE)
}
\arguments{
\item{data}{The data to check}

\item{level}{character_array the level of the data to check}

\item{localised}{logical to check localised data or not, defaults to
TRUE.}
}
\description{
Checks that processed data columns date, cases_new, cases_total,
deaths_new, deaths_total and that region level have the correct types.
}
\seealso{
Functions used for testing data is cleaned and processed correctly
\code{\link{expect_clean_cols}()},
\code{\link{expect_columns_contain_data}()},
\code{\link{test_cleaning}()},
\code{\link{test_download}()},
\code{\link{test_processing}()},
\code{\link{test_return}()}
}
\concept{tests}
