% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-methods.R
\name{general_init}
\alias{general_init}
\title{initialize function used by all \code{Country} class objects.}
\usage{
general_init(
  self,
  level = "1",
  totals = FALSE,
  localise = TRUE,
  verbose = TRUE,
  steps = FALSE
)
}
\arguments{
\item{self}{The specific class object to attach values}

\item{level}{A character string indicating the target administrative level
of the data with the default being "1". Currently supported options are
level 1 ("1) and level 2 ("2"). Use \code{get_available_datasets} for supported
options by dataset.}

\item{totals}{Logical, defaults to FALSE. If TRUE, returns totalled
data per region up to today's date. If FALSE, returns the full dataset
stratified by date and region.}

\item{localise}{Logical, defaults to TRUE. Should region names be localised.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{steps}{Logical, defaults to FALSE. Should all processing and cleaning
steps be kept and output in a list.}
}
\description{
Set up the country class with attributes set to input
parameters. Should only be called by a \code{Country} class object.
}
