% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_afghan_regional_cases.R
\name{get_afghan_regional_cases}
\alias{get_afghan_regional_cases}
\title{Afghan Provincial Daily Case Counts}
\usage{
get_afghan_regional_cases()
}
\value{
A dataframe of daily Afghan provincial cases and deaths
}
\description{
Data from HDX https://data.humdata.org/dataset/afghanistan-covid-19-statistics-per-province
The data is stored in a google sheet, which is read as a csv.
}
\examples{

## Code
get_afghan_regional_cases()
}
\author{
Flavio Finger @ffinger
}
