% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_countries.R
\name{get_available_datasets}
\alias{get_available_datasets}
\title{Get the countries data is available for}
\usage{
get_available_datasets()
}
\value{
A list of available countries and the region level
data is available for
}
\description{
Show what countries have what level data available.
The function searches the environment for R6 class objects and
extracts the country name and what level it has from the object.
}
\examples{
\dontrun{
get_avaliable_datasets()
}

}
