% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/belgium.R
\name{get_belgium_regional_cases_only_level_1}
\alias{get_belgium_regional_cases_only_level_1}
\title{Belgian Regional Daily COVID-19 Count Data - Regions Only}
\usage{
get_belgium_regional_cases_only_level_1()
}
\value{
A data frame of COVID cases by Region in Belgium, stratified by region, ready to be used by \code{get_regional_data()}.
}
\description{
Fetches daily COVID-19 data from Sciensano, the Belgian Institute for Health.
Data is available at \url{https://epistat.wiv-isp.be/covid/}.
It is loaded and then sanitised.
}
