% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/belgium.R
\name{get_belgium_regional_cases_with_level_2}
\alias{get_belgium_regional_cases_with_level_2}
\title{Belgian Provincial Daily COVID-19 Count Data - Regions and Provinces}
\usage{
get_belgium_regional_cases_with_level_2()
}
\value{
A data frame of COVID cases by province in Belgium, stratified by province, ready to be used by get_regional_data().
}
\description{
Fetches daily COVID data from Sciensano, the Belgian Institute for Health.
Data is available at \url{https://epistat.wiv-isp.be/covid/}.
It is then loaded and sanitised.
}
