% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_regional_cases.R
\name{get_brazil_regional_cases}
\alias{get_brazil_regional_cases}
\title{Brazil regional cases and deaths}
\usage{
get_brazil_regional_cases()
}
\description{
Extracts regional case and death counts for Brazil.
Data available on Github, curated by Wesley Cota: DOI 10.1590/SciELOPreprints.362
\href{https://github.com/wcota/covid19br}{Source}
}
