% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_canada_regional_cases.R
\name{get_canada_regional_cases}
\alias{get_canada_regional_cases}
\title{Fetch daily COVID cases by province for Canada}
\usage{
get_canada_regional_cases(out = "timeseries")
}
\arguments{
\item{out}{Character String specifying output data format "timeseries", "total", default: "timeseries".}
}
\value{
A data.frame of COVID cases by province in Canada
}
\description{
Fetches daily COVID cases, and deaths by province collated by Provincial Canadian Health Authorities
Data is available at https://www.canada.ca/en/public-health/services/diseases/2019-novel-coronavirus-infection.html
}
\examples{

\dontrun{

 get_canada_regional_cases(out = 'timeseries')

}
}
