% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada.R
\name{get_canada_regional_cases}
\alias{get_canada_regional_cases}
\title{Canadian Regional Daily COVID-19 Count Data - Provinces}
\usage{
get_canada_regional_cases()
}
\value{
A data frame of COVID cases by province in Canada, ready to be used by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for Canada, stratified by province.
Data available at  \url{https://health-infobase.canada.ca}.
It is loaded and then sanitised.
}
