% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colombia.R
\name{get_colombia_regional_cases}
\alias{get_colombia_regional_cases}
\title{Colombian Regional Daily COVID-19 Count Data - Department}
\usage{
get_colombia_regional_cases()
}
\value{
A data frame of COVID cases by department in Colombia, ready to be used by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for Colombia, stratified by departamento.
Data available at  \url{https://github.com/danielcs88/colombia_covid-19}.
It is loaded and then cleaned.
}
