% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national.R
\name{get_ecdc_cases}
\alias{get_ecdc_cases}
\title{ECDC International Case Counts: works within get_national_data}
\usage{
get_ecdc_cases()
}
\value{
A dataframe of International case counts published by ECDC.
}
\description{
ECDC International Case Counts: works within get_national_data
}
\author{
Sam Abbott @seabbs

D. van Muijen @dmuijen

Kath Sherratt @kathsherratt

Haze Lee @hazealign
}
