% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_germany_regional_cases.R
\name{get_germany_regional_cases}
\alias{get_germany_regional_cases}
\title{Get German daily cases by Bundeslander}
\usage{
get_germany_regional_cases()
}
\value{
A dataframe of case counts in German regions
}
\description{
Fetches COVID case counts by region in Germany.
This data was collated by Jan-Philip Gehrcke (gh: jgehrcke) and is available at:
https://github.com/jgehrcke/covid-19-germany-gae
}
\examples{


\dontrun{
country <- rnaturalearth::ne_countries(scale="large",
                                        country = "Germany",
                                        returnclass = 'sf')

regions <- rnaturalearth::ne_states("Germany", returnclass = "sf")

data <- get_germany_regional_cases() \%>\%
  dplyr::filter(date == max(date))

regions_with_data <- regions \%>\%
  dplyr::left_join(data,
                  by = c("iso_3166_2" = "region_code"))

ggplot2::ggplot(regions_with_data) +
  ggplot2::geom_sf(ggplot2::aes(fill = cases))

}
}
