% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_germany_regional_cases.R
\name{get_germany_regional_cases}
\alias{get_germany_regional_cases}
\title{Get German daily cases by Bundesländer}
\usage{
get_germany_regional_cases()
}
\value{
A dataframe of case counts in German regions
}
\description{
Get German daily cases by Bundesländer
}
\examples{

}
