% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_germany_regional_cases.R
\name{get_germany_regional_cases}
\alias{get_germany_regional_cases}
\title{Get German daily cases by Bundeslander or Landkreis}
\usage{
get_germany_regional_cases(geography = "states")
}
\value{
A dataframe of case and death counts in German regions
}
\description{
Fetches COVID case counts by region in Germany.
This data is sourced from the Robert Koch Institute:
https://npgeo-corona-npgeo-de.hub.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0
}
\examples{


\dontrun{
country <- rnaturalearth::ne_countries(scale="large",
                                        country = "Germany",
                                        returnclass = 'sf')

regions <- rnaturalearth::ne_states("Germany", returnclass = "sf")

data <- get_germany_regional_cases() \%>\%
  dplyr::filter(date == max(date))

regions_with_data <- regions \%>\%
  dplyr::left_join(data,
                  by = c("name" = "state"))

ggplot2::ggplot(regions_with_data) +
  ggplot2::geom_sf(ggplot2::aes(fill = cases))

}

}
