% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germany.R
\name{get_germany_regional_cases_only_level_1}
\alias{get_germany_regional_cases_only_level_1}
\title{German Regional Daily COVID-19 Count Data - Bundesland}
\usage{
get_germany_regional_cases_only_level_1()
}
\value{
A data frame of COVID cases by Bundesland in Germany, ready to be used by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for Germany, stratified by Bundesland.
Data available at  \url{https://npgeo-corona-npgeo-de.hub.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0}.
It is loaded and then sanitised.
}
