% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germany.R
\name{get_germany_regional_cases_with_level_2}
\alias{get_germany_regional_cases_with_level_2}
\title{German Regional Daily COVID-19 Count Data - Landkreis}
\usage{
get_germany_regional_cases_with_level_2()
}
\value{
A data.frame of COVID cases by Landkreis in Germany, ready to be used by get_regional_data().
}
\description{
Extracts daily COVID-19 data for Germany, stratified by Landkreis.
Data available at  \url{https://opendata.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0.csv}.
It is loaded and then sanitised.
}
