% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/india.R
\name{get_india_regional_cases}
\alias{get_india_regional_cases}
\title{Get India daily COVID data}
\usage{
get_india_regional_cases()
}
\value{
A dataframe of daily India data to be further processed by \code{\link[=get_regional_covid_data]{get_regional_covid_data()}}.
}
\description{
Get Indian cases, deaths and recoveries by state
}
