% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/india.R
\name{get_india_regional_cases}
\alias{get_india_regional_cases}
\title{Indian Regional Daily COVID-19 Count Data - State}
\usage{
get_india_regional_cases()
}
\value{
A dataframe of daily India data to be further processed by \code{\link[=get_regional_data]{get_regional_data()}}.
}
\description{
Extracts daily COVID-19 data for India, stratified by State
Data available at  \url{https://api.covid19india.org/csv/latest/state_wise_daily.csv}.
It is loaded and then sanitised.
}
