% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_covidregionaldata.R
\name{get_info_covidregionaldata}
\alias{get_info_covidregionaldata}
\title{Get available datasets}
\usage{
get_info_covidregionaldata()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{get_available_datasets()} instead.
}
\keyword{internal}
