% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_international_linelist.R
\name{get_international_linelist}
\alias{get_international_linelist}
\title{Get International Linelist Data}
\usage{
get_international_linelist(countries = NULL, cities = NULL,
  provinces = NULL, clean = TRUE)
}
\arguments{
\item{countries}{Character vector identifying the countries to extract data for.}

\item{cities}{Character vector identifying the cities to extract data for.}

\item{provinces}{Character vector identifying the provinces to extract data for.}

\item{clean}{Logical, defaults to \code{TRUE}. Should the data returned be cleaned for use.}
}
\value{
A linelist of case data
}
\description{
This function downloads the latest non-Hubei linelist. It uses \code{memoise} to cache the
results locally. To clear the cache and redownload the data use \code{reset_cache}. The cache will be located
in the directory at which the function is run. As this linelist is experiencing a high user demand it may not always be available.
To account for this we keep a cache in the \code{NCoVUtils} GitHub repo which this function will fall back to with a warning if the source
cannot be downloaded.
}
\examples{

get_international_linelist
}
\author{
Sam Abbott \href{mailto:sam.abbott@lshtm.ac.uk}{sam.abbott@lshtm.ac.uk}
}
