% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_italy_regional_cases.R
\name{get_italy_regional_cases}
\alias{get_italy_regional_cases}
\title{Italian Regional Case Counts}
\usage{
get_italy_regional_cases()
}
\value{
A dataframe of Italian regional case counts.
}
\description{
Fetches COVID data by region. Data is collated by the Italian Department of Civil Protection
and is available on github: https://github.com/pcm-dpc/COVID-19
}
\examples{

## Code
get_italy_regional_cases
}
