% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_italy_regional_cases.R
\name{get_italy_regional_cases}
\alias{get_italy_regional_cases}
\title{Italian Regional Covid-19 Data}
\usage{
get_italy_regional_cases()
}
\value{
A data frame of Italian regional case counts ready to be used by \code{get_regional_covid_data()}
}
\description{
Fetches COVID data by region. Data is collated by the Italian Department of Civil Protection
and is available on github: https://github.com/pcm-dpc/COVID-19. Cleans and sanitises ready for further use.
}
