% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy.R
\name{get_italy_regional_cases}
\alias{get_italy_regional_cases}
\title{Regional Daily COVID-19 Count Data}
\usage{
get_italy_regional_cases()
}
\value{
A data frame of daily COVID cases for Italy by region, to be further processed by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for Italy, stratified by Region.
Data available at  \url{https://raw.githubusercontent.com/pcm-dpc/COVID-19/master/dati-regioni/dpc-covid19-ita-regioni.csv}.
It is loaded and then sanitised.
}
