% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_japan_regional_cases.R
\name{get_japan_regional_cases}
\alias{get_japan_regional_cases}
\title{Japan regional cases, daily}
\usage{
get_japan_regional_cases()
}
\description{
Extracts regional case counts for Japan.
\href{https://en.wikipedia.org/wiki/2020_coronavirus_pandemic_in_Japan}{Source}.
}
\examples{
get_japan_regional_cases()

\dontrun{

library(ggplot2)
library(dplyr)

 # Mapping: set up
 data <- get_japan_regional_cases()

 regions <- rnaturalearth::ne_states("Japan", returnclass = "sf")

 regions_with_data <- regions \%>\%
   dplyr::left_join(data, by = c("name" = "region")) \%>\%
   dplyr::mutate(cases =  ifelse(is.na(cases), 0, cases)

   # Map inset: Okinawa
   jp_okinawa <- dplyr::filter(regions_with_data, name == "Okinawa")

   jp_okinawa <- ggplot2::ggplot(jp_okinawa) +
   geom_sf(aes(fill = cases)) +  coord_sf(datum = NA) +
   xlab(jp_okinawa$region) +
   theme_bw() +
   theme(legend.position = "none")

   ## Map: mainland with insets
   jp_main <- dplyr::filter(regions_with_data, name != "Okinawa") \%>\%
   ggplot() +
   geom_sf(aes(fill = cases)) +
   coord_sf(crs = sf::st_crs(4326), xlim = c(127, 146), ylim = c(29, 46)) +
   theme_bw()

   jp_main +
   annotation_custom(
   grob = ggplotGrob(jp_okinawa), xmin = 140,  xmax = 146,  ymin = 24,  ymax = 37)

}
}
