% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHU.R
\name{get_JHU_data}
\alias{get_JHU_data}
\title{Download and clean data from JHU}
\usage{
get_JHU_data()
}
\value{
A data frame of COVID cases for all countris listed by JHU, by country and region if avaliable
}
\description{
Downloads daily covid data from JHU
(https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data)
and formats the data
}
