% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linelist.R
\name{get_linelist}
\alias{get_linelist}
\title{Get patient line list data}
\source{
\url{https://github.com/beoutbreakprepared/nCoV2019}
}
\usage{
get_linelist(clean = TRUE, report_delay_only = FALSE)
}
\arguments{
\item{clean}{Logical, defaults to \code{TRUE}.
Should the data returned be cleaned for use.}

\item{report_delay_only}{Logical, defaults to \code{FALSE}.
Should only certain variables (id, country, onset date, days' delay),
and observations (patients with a report delay) be returned}
}
\value{
A line list of reported cases of COVID-19
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Provides a public international patient line list from January 2020 to June
2020.

This version of the line list has stopped updating. The new version of the
line list is free but requires a login.

See: https://global.health/
}
\examples{
\dontrun{
# Get the complete linelist
get_linelist()

# Return the report delay only
get_linelist(report_delay_only = TRUE)
}

}
\keyword{internal}
