% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lithuania.R
\name{get_lithuania_regional_cases_with_level_2}
\alias{get_lithuania_regional_cases_with_level_2}
\title{Lithuanian Daily COVID-19 Count Data - Municipalities}
\usage{
get_lithuania_regional_cases_with_level_2()
}
\value{
A data.frame of COVID cases by municipality in Lithuania, ready to be used by get_regional_data().
}
\description{
Extracts daily COVID-19 data for Lithuania, by municipality.
Data available at \url{https://opendata.arcgis.com/datasets/45b76303953d40e2996a3da255bf8fe8_0}.
It is loaded and then sanitised.
}
