% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UK.R, R/shared-methods.R
\name{Uk}
\alias{Uk}
\alias{DataClass}
\title{United Kingdom Class for downloading, cleaning and processing notification
data.}
\source{
https://coronavirus.data.gov.uk/details/download #nolint
}
\description{
Country specific information for downloading, cleaning
and processing covid-19 region data for an example Country.

Acts as parent class for individual country objects,
allowing them to access general methods.
}
\details{
Inherits from \code{DataClass}

All countries have shared methods for extracting region codes,
downloading, processing, and returning data.
}
\examples{
\dontrun{
region <- Italy$new(verbose = TRUE, steps = TRUE)
region$download()
region$clean()
region$process()
region$return()
}
}
\section{Super class}{
\code{\link[covidregionaldata:DataClass]{covidregionaldata::DataClass}} -> \code{Uk}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level_1_region}}{the level 1 region name.}

\item{\code{level_2_region}}{the level 2 region name}

\item{\code{data_url}}{link to raw data}

\item{\code{source_data_cols}}{existing columns within the raw data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_region_codes}{\code{Uk$get_region_codes()}}
\item \href{#method-download}{\code{Uk$download()}}
\item \href{#method-clean}{\code{Uk$clean()}}
\item \href{#method-clean_level_1}{\code{Uk$clean_level_1()}}
\item \href{#method-clean_level_2}{\code{Uk$clean_level_2()}}
\item \href{#method-process}{\code{Uk$process()}}
\item \href{#method-new}{\code{Uk$new()}}
\item \href{#method-clone}{\code{Uk$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="DataClass" data-id="return">}\href{../../covidregionaldata/html/DataClass.html#method-return}{\code{covidregionaldata::DataClass$return()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_region_codes"></a>}}
\if{latex}{\out{\hypertarget{method-get_region_codes}{}}}
\subsection{Method \code{get_region_codes()}}{
Specific function for getting region codes for UK .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$get_region_codes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
UK specific download function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$download()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\if{latex}{\out{\hypertarget{method-clean}{}}}
\subsection{Method \code{clean()}}{
UK specific cleaning, directs to level 1 or level 2
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_level_1"></a>}}
\if{latex}{\out{\hypertarget{method-clean_level_1}{}}}
\subsection{Method \code{clean_level_1()}}{
UK Specific Region Level Data Cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$clean_level_1()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_level_2"></a>}}
\if{latex}{\out{\hypertarget{method-clean_level_2}{}}}
\subsection{Method \code{clean_level_2()}}{
UK Specific Level 2 Data Cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$clean_level_2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process"></a>}}
\if{latex}{\out{\hypertarget{method-process}{}}}
\subsection{Method \code{process()}}{
UK specific function to processes regional data.
Calls the parent process function but for level two regions does some
column renaming.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$process()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Specific initalize function for UK providing extra
arguments specific to the UK
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$new(
  level = "1",
  totals = FALSE,
  localise = TRUE,
  verbose = TRUE,
  steps = FALSE,
  nhsregions = FALSE,
  release_date = NULL,
  resolution = "utla"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{A character string indicating the target administrative
level of the data with the default being "1". Currently supported
options are level 1 ("1) and level 2 ("2").
Use \code{get_available_datasets} for supported options by dataset.}

\item{\code{totals}}{Logical, defaults to FALSE. If TRUE, returns totalled
data per region up to today's date. If FALSE, returns the full dataset
stratified by date and region.}

\item{\code{localise}}{Logical, defaults to TRUE. Should region names be
localised.}

\item{\code{verbose}}{Logical, defaults to \code{TRUE}. Should verbose processing
messages and warnings be returned.}

\item{\code{steps}}{Logical, defaults to FALSE. Should all processing and
cleaning steps be kept and output in a list.
#' @export}

\item{\code{nhsregions}}{Return subnational English regions using NHS region
boundaries instead of PHE boundaries.}

\item{\code{release_date}}{Date data was released. Default is to extract
latest release. Dates should be in the format "yyyy-mm-dd".}

\item{\code{resolution}}{"utla" (default) or "ltla", depending on which
geographical resolution is preferred}

\item{\code{self}}{The specific class object to attach values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uk$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{country}}{name of country to fetch data for}

\item{\code{region}}{data frame for requested region}

\item{\code{data_url}}{link to raw data}

\item{\code{level}}{target region level}

\item{\code{totals}}{Boolean. If TRUE, returns totalled data per region
up to today's date.}

\item{\code{localise}}{Boolean. Should region names be localised.}

\item{\code{verbose}}{Boolean. Display information at various stages.}

\item{\code{steps}}{Boolean. Keep data from each processing step.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_region_codes}{\code{DataClass$get_region_codes()}}
\item \href{#method-download}{\code{DataClass$download()}}
\item \href{#method-process}{\code{DataClass$process()}}
\item \href{#method-return}{\code{DataClass$return()}}
\item \href{#method-clone}{\code{DataClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_region_codes"></a>}}
\if{latex}{\out{\hypertarget{method-get_region_codes}{}}}
\subsection{Method \code{get_region_codes()}}{
General function for getting region codes for given region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$get_region_codes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
General function for downloading raw data.
Shared regional dataset processing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$download()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process"></a>}}
\if{latex}{\out{\hypertarget{method-process}{}}}
\subsection{Method \code{process()}}{
General function to processes regional data.
Dynamically works for level 1 and level 2 regions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$process()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return"></a>}}
\if{latex}{\out{\hypertarget{method-return}{}}}
\subsection{Method \code{return()}}{
Optional region specific return changes.
Designed to be called after \code{process}. For most datasets a
custom method should not be needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$return()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
