% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-methods.R
\name{dataClass}
\alias{dataClass}
\title{R6 Class containing non-country specific methods}
\description{
Acts as parent class for individual country objects,
allowing them to access general methods.
}
\details{
All countries have shared methods for extracting region codes,
downloading, processing, and returning data.
}
\author{
Sam Abbott
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{country}}{name of country to fetch data for}

\item{\code{region}}{data frame for requested region}

\item{\code{data_url}}{link to raw data}

\item{\code{level}}{target region level}

\item{\code{totals}}{Boolean. If TRUE, returns totalled data per region up to today's date.}

\item{\code{localise}}{Boolean. Should region names be localised.}

\item{\code{verbose}}{Boolean. Display information at various stages.}

\item{\code{steps}}{Boolean. Keep data from each processing step.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_region_codes}{\code{dataClass$get_region_codes()}}
\item \href{#method-download_data}{\code{dataClass$download_data()}}
\item \href{#method-process_regional}{\code{dataClass$process_regional()}}
\item \href{#method-return_regional}{\code{dataClass$return_regional()}}
\item \href{#method-clone}{\code{dataClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_region_codes"></a>}}
\if{latex}{\out{\hypertarget{method-get_region_codes}{}}}
\subsection{Method \code{get_region_codes()}}{
General function for getting region codes for given region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataClass$get_region_codes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_data"></a>}}
\if{latex}{\out{\hypertarget{method-download_data}{}}}
\subsection{Method \code{download_data()}}{
General function for downloading raw data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataClass$download_data(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{pass additional arguments
Shared regional dataset processing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process_regional"></a>}}
\if{latex}{\out{\hypertarget{method-process_regional}{}}}
\subsection{Method \code{process_regional()}}{
General function to processes regional data.
Dynamically works for level 1 and level 2 regions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataClass$process_regional(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{pass additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_regional"></a>}}
\if{latex}{\out{\hypertarget{method-return_regional}{}}}
\subsection{Method \code{return_regional()}}{
Optional region specific return changes. Designed to be called
after \code{process_regional}. For most datasets a custom method should not be
needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataClass$return_regional()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
