% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_codes.R
\name{get_region_codes}
\alias{get_region_codes}
\title{Get a table of region codes for a specified country}
\usage{
get_region_codes(country)
}
\arguments{
\item{country}{a string with a country specified}
}
\value{
a tibble of regions and their corresponding region codes
}
\description{
Get a table of region codes for a specified country
}
