% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/russia.R
\name{get_russia_regional_cases}
\alias{get_russia_regional_cases}
\title{Russian Regional Daily COVID-19 Count Data - Region}
\usage{
get_russia_regional_cases()
}
\value{
A data frame of daily COVID cases for Russia by region, to be further processed by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for Russia, stratified by Region.
Data available at
\url{https://raw.githubusercontent.com/grwlf/COVID-19_plus_Russia/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_RU.csv}.
It is loaded and then sanitised.
}
