% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spain_regional_cases.R
\name{get_spain_regional_cases}
\alias{get_spain_regional_cases}
\title{Spain regional cases daily}
\usage{
get_spain_regional_cases(dataset = "cases_provincial")
}
\arguments{
\item{dataset}{Character String specifying dataset: "cases_provincial", "hospitalisation_provincial", "icu_provincial", "mortality_provincial", "recovered_provincial", "all". Default: "cases_provincial".}
}
\value{
A dataframe of specified Covid data. If dataset = "all", a dataframe with all variables (cases, hospitalisation, ICU, mortality, recovered)
}
\description{
Extract regional spanish case counts.
\href{https://covid19.isciii.es/}{Source}
}
\examples{

\dontrun{

# Regional map

library(sf)
library(rnaturalearth)
library(ggplot2)

data <- get_spain_regional_cases()

regions <- rnaturalearth::ne_states("Spain", returnclass = "sf")
regions_with_data <- regions \%>\%
  mutate(region_shortcode = stringr::str_remove_all(region_cod, "^ES\\."), # Match codes to dataset
         region_shortcode = stringr::str_replace(region_shortcode, "^PM$", "IB"), # Baleares
         region_shortcode = stringr::str_replace(region_shortcode, "^MU$", "MC"), # Murcia
         region_shortcode = stringr::str_replace(region_shortcode, "^NA$", "NC"), # Navarra
         region_shortcode = stringr::str_replace(region_shortcode, "^LO$", "RI"), # La Rioja
         region_shortcode = ifelse(region == "Melilla", "ME", region_shortcode) # Melilla
  ) \%>\%
  group_by(region_shortcode) \%>\%
  summarise(subregions = n(), do_union = TRUE) \%>\%
  left_join(data, by = c("region_shortcode" = "region"))

# Map: Spain mainland + Baleares
es_mainland <- ggplot2::ggplot(regions_with_data) +
  geom_sf(ggplot2::aes(fill = cases)) +
  coord_sf(crs = st_crs(4326), xlim = c(-11, 4), ylim = c(35, 44))+
  theme_bw()

# Map inset: Canarias
es_canarias <- dplyr::filter(regions_with_data, region == "Canarias")
es_canarias <- ggplot2::ggplot(es_canarias) +
  geom_sf(aes(fill = cases)) +  coord_sf(datum = NA) +
  xlab(es_canarias$region) +
  theme_bw() +  theme(legend.position = "none")

# Map: mainland with insets
es_mainland +
  annotation_custom(
    grob = ggplotGrob(es_canarias),  xmin = -11,  xmax = -7,  ymin = 33.5,  ymax = 37.5)


}

}
