% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_cases.R
\name{get_total_cases}
\alias{get_total_cases}
\title{Get Total Cases by Country}
\usage{
get_total_cases(source = "WHO")
}
\arguments{
\item{source}{Character String specifying data source for total cases "WHO", "ECDC", default: "WHO".}
}
\value{
Returns a data frame of variables along with their most recent value.
}
\description{
Get Total Cases by Country
}
