% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{get_uk_data}
\alias{get_uk_data}
\title{Get UK data - helper function to get data for a single valid area type}
\usage{
get_uk_data(filters, progress_bar = FALSE)
}
\arguments{
\item{filters}{Query filters for UK data e.g. "areaType=nation"}

\item{progress_bar}{Display a progress bar, useful for level 2 which takes a while to load}
}
\value{
A dataframe with all variables available in public UK data
}
\description{
UK data download helper. Code adapted from PHE Coronavirus API for R:
available at https://github.com/publichealthengland/coronavirus-dashboard-api-R-sdk
}
