% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uk_nhs_region_cases.R
\name{get_uk_nhs_region_cases}
\alias{get_uk_nhs_region_cases}
\title{Get UK daily cases by NHS region}
\usage{
get_uk_nhs_region_cases()
}
\value{
A dataframe of case counts in English NHS regions
}
\description{
Get UK regional cases.
}
\examples{
get_uk_nhs_region_cases

\dontrun{
uk_shp <- readRDS("uk_shp.rds")
cases <- covidregionaldata::get_uk_regional_cases()
cases <- cases[1:23,]
uk_shp \%>\%
dplyr::full_join(cases, by = "region") \%>\%
 ggplot2::ggplot(ggplot2::aes(fill = cases)) + ggplot2::geom_sf()
}
}
