% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{get_uk_regional_cases_only_level_1}
\alias{get_uk_regional_cases_only_level_1}
\title{UK Regional Daily COVID-19 Count Data - Region}
\usage{
get_uk_regional_cases_only_level_1()
}
\value{
A data frame of daily COVID cases for the UK by region, to be further processed by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for the UK, stratified by region and nation.
Data source:
}
