% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{get_uk_regional_cases_with_level_2}
\alias{get_uk_regional_cases_with_level_2}
\title{UK Regional Daily COVID-19 Count Data - UTLA}
\usage{
get_uk_regional_cases_with_level_2()
}
\value{
A data frame of daily COVID cases for the UK by region, to be further processed by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for the UK, stratified by Upper Tier Local Authority
Data source:
}
