% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_US_regional_cases.R
\name{get_US_regional_cases}
\alias{get_US_regional_cases}
\title{Fetch daily COVID cases by state for United States of America}
\usage{
get_US_regional_cases()
}
\value{

}
\description{
Fetches daily COVID cases by state collated by the
John Hopkins Centre for Systems Science and Engineering (JHU CSSE)
Data is available at https://github.com/CSSEGISandData/COVID-19
}
\examples{

}
