% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_US_regional_cases.R
\name{get_US_regional_cases}
\alias{get_US_regional_cases}
\title{Fetch daily COVID cases by state for United States of America}
\usage{
get_US_regional_cases()
}
\value{

}
\description{
Fetch daily COVID cases by state for United States of America
}
\examples{

}
