% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa.R
\name{get_us_regional_cases}
\alias{get_us_regional_cases}
\title{Fetch daily COVID cases by state for United States of America}
\usage{
get_us_regional_cases(level = "state", out = "timeseries")
}
\arguments{
\item{level}{Character String specifying admin level "state", "county", default: "state".}

\item{out}{Character String specifying output data format "timeseries", "total", default: "timeseries".}
}
\value{
A data.frame of COVID cases by region in the US
}
\description{
Fetches daily COVID cases by state or county collated by the
New York Times
Data is available at https://github.com/nytimes/covid-19-data
}
\examples{
get_us_regional_cases(level = 'state', out = 'timeseries')

\dontrun{

 regions <- rnaturalearth::ne_states("United States of America", returnclass = "sf")

 cases <- get_us_regional_cases(out = 'total')

 regions \%>\% dplyr::filter(name \%in\% cases$state) \%>\%
 ggplot2::ggplot() + ggplot2::geom_sf()

}
}
