% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa.R
\name{get_us_regional_cases_only_level_1}
\alias{get_us_regional_cases_only_level_1}
\title{US Regional Daily COVID-19 Count Data - States}
\usage{
get_us_regional_cases_only_level_1()
}
\value{
A data frame of daily COVID cases for the US by state, to be further processed by \code{get_regional_data()}.
}
\description{
Extracts daily COVID-19 data for the USA, stratified by state.
Data available at \url{https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-states.csv}.
It is loaded and then sanitised.
}
