% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national.R
\name{get_who_cases}
\alias{get_who_cases}
\title{Download the most recent WHO case data}
\usage{
get_who_cases()
}
\value{
A tibble of all WHO data by date
}
\description{
Downloads the latest WHO case data.
}
