% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Italy.R
\name{Italy}
\alias{Italy}
\title{R6 Class containing country specific attributes and methods}
\description{
Country specific information for downloading, cleaning
and processing covid-19 region data for Italy.
}
\details{
Inherits from \code{dataClass}
}
\section{Super class}{
\code{\link[covidregionaldata:dataClass]{covidregionaldata::dataClass}} -> \code{italy}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level_1_region}}{the level 1 region name.}

\item{\code{level_2_region}}{the level 2 region name.}

\item{\code{data_url}}{link to raw data}

\item{\code{source_data_cols}}{existing columns within the raw data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clean}{\code{Italy$clean()}}
\item \href{#method-new}{\code{Italy$new()}}
\item \href{#method-clone}{\code{Italy$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="download">}\href{../../covidregionaldata/html/dataClass.html#method-download}{\code{covidregionaldata::dataClass$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="get_region_codes">}\href{../../covidregionaldata/html/dataClass.html#method-get_region_codes}{\code{covidregionaldata::dataClass$get_region_codes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="process">}\href{../../covidregionaldata/html/dataClass.html#method-process}{\code{covidregionaldata::dataClass$process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="return">}\href{../../covidregionaldata/html/dataClass.html#method-return}{\code{covidregionaldata::dataClass$return()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean"></a>}}
\if{latex}{\out{\hypertarget{method-clean}{}}}
\subsection{Method \code{clean()}}{
Italy specific state level data cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Italy$clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Set up the country class with attributes set to input
parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Italy$new(level, totals, localise, verbose, steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The region level for the data}

\item{\code{totals}}{Boolean. If TRUE, returns totalled data per region up to
today's date.}

\item{\code{localise}}{Boolean. Should region names be localised.}

\item{\code{verbose}}{Boolean. Display information at various stages.}

\item{\code{steps}}{Boolean. Keep data from each processing step.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Italy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
