% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexico.R
\name{Mexico}
\alias{Mexico}
\title{R6 Class containing country specific attributes and methods}
\description{
Country specific information for downloading, cleaning
and processing covid-19 region data for Mexico.
}
\details{
Inherits from \code{dataClass}
}
\author{
Sam Abbott
}
\section{Super class}{
\code{\link[covidregionaldata:dataClass]{covidregionaldata::dataClass}} -> \code{mexico}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level_1_region}}{the level 1 region name}

\item{\code{level_2_region}}{the level 2 region name.}

\item{\code{data_url}}{link to raw data}

\item{\code{source_data_cols}}{existing columns within the raw data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-download_data}{\code{Mexico$download_data()}}
\item \href{#method-clean_regional}{\code{Mexico$clean_regional()}}
\item \href{#method-clean_regional_level_1}{\code{Mexico$clean_regional_level_1()}}
\item \href{#method-clean_regional_level_2}{\code{Mexico$clean_regional_level_2()}}
\item \href{#method-new}{\code{Mexico$new()}}
\item \href{#method-clone}{\code{Mexico$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="get_region_codes">}\href{../../covidregionaldata/html/dataClass.html#method-get_region_codes}{\code{covidregionaldata::dataClass$get_region_codes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="process_regional">}\href{../../covidregionaldata/html/dataClass.html#method-process_regional}{\code{covidregionaldata::dataClass$process_regional()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="return_regional">}\href{../../covidregionaldata/html/dataClass.html#method-return_regional}{\code{covidregionaldata::dataClass$return_regional()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_data"></a>}}
\if{latex}{\out{\hypertarget{method-download_data}{}}}
\subsection{Method \code{download_data()}}{
Data download function for Mexico data. This replaces the
generic download_data function in \code{dataClass}. To get the latest data
use a PHP script from the website.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mexico$download_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_regional"></a>}}
\if{latex}{\out{\hypertarget{method-clean_regional}{}}}
\subsection{Method \code{clean_regional()}}{
directs to either level 1 or level 2 processing based on
request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mexico$clean_regional(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{pass additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_regional_level_1"></a>}}
\if{latex}{\out{\hypertarget{method-clean_regional_level_1}{}}}
\subsection{Method \code{clean_regional_level_1()}}{
Mexico Specific Estados Level Data Cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mexico$clean_regional_level_1(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{pass additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_regional_level_2"></a>}}
\if{latex}{\out{\hypertarget{method-clean_regional_level_2}{}}}
\subsection{Method \code{clean_regional_level_2()}}{
Mexico Specific Municipality Level Data Cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mexico$clean_regional_level_2(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{pass additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Set up the country class with attributes set to input
parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mexico$new(level, totals, localise, verbose, steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The region level for the data}

\item{\code{totals}}{Boolean. If TRUE, returns totalled data per region up to
today's date.}

\item{\code{localise}}{Boolean. Should region names be localised.}

\item{\code{verbose}}{Boolean. Display information at various stages.}

\item{\code{steps}}{Boolean. Keep data from each processing step.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mexico$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
