% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mexico_codes}
\alias{mexico_codes}
\title{Region Codes for Mexico Dataset.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2489 rows and 4 columns.
}
\usage{
mexico_codes
}
\value{
A nested tibble of region codes and related information.
}
\description{
Details of the region codes used for the Mexico dataset.
}
\keyword{datasets}
