% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Country_template.R
\name{NameOfCountry}
\alias{NameOfCountry}
\title{R6 Class containing country specific attributes and methods}
\description{
Country specific information for downloading, cleaning
and processing covid-19 region data for \emph{Country-name}
}
\details{
Inherits from \code{dataClass}
}
\section{Super class}{
\code{\link[covidregionaldata:dataClass]{covidregionaldata::dataClass}} -> \code{country_name}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level_1_region}}{the level 1 region name.}

\item{\code{level_2_region}}{the level 2 region name.}

\item{\code{data_url}}{link to raw data}

\item{\code{source_data_cols}}{existing columns within the raw data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clean_regional}{\code{NameOfCountry$clean_regional()}}
\item \href{#method-new}{\code{NameOfCountry$new()}}
\item \href{#method-clone}{\code{NameOfCountry$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="download_data">}\href{../../covidregionaldata/html/dataClass.html#method-download_data}{\code{covidregionaldata::dataClass$download_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="get_region_codes">}\href{../../covidregionaldata/html/dataClass.html#method-get_region_codes}{\code{covidregionaldata::dataClass$get_region_codes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="process_regional">}\href{../../covidregionaldata/html/dataClass.html#method-process_regional}{\code{covidregionaldata::dataClass$process_regional()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="covidregionaldata" data-topic="dataClass" data-id="return_regional">}\href{../../covidregionaldata/html/dataClass.html#method-return_regional}{\code{covidregionaldata::dataClass$return_regional()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_regional"></a>}}
\if{latex}{\out{\hypertarget{method-clean_regional}{}}}
\subsection{Method \code{clean_regional()}}{
\emph{Country-name} specific state level data cleaning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameOfCountry$clean_regional(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{pass additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Set up the country class with attributes set to input
parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameOfCountry$new(level, totals, localise, verbose, steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The region level for the data}

\item{\code{totals}}{Boolean. If TRUE, returns totalled data per region up
to today's date.}

\item{\code{localise}}{Boolean. Should region names be localised.}

\item{\code{verbose}}{Boolean. Display information at various stages.}

\item{\code{steps}}{Boolean. Keep data from each processing step.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameOfCountry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
