% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_default_processing_fns}
\alias{run_default_processing_fns}
\title{Default processing steps to run}
\usage{
run_default_processing_fns(data)
}
\arguments{
\item{data}{A data table}
}
\description{
The default processing steps to which are always run. Runs on
clean data
}
\seealso{
Functions used in the processing pipeline
\code{\link{process_internal}()},
\code{\link{run_optional_processing_fns}()}
}
\concept{processing}
\concept{utility}
