% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_process_steps}
\alias{run_process_steps}
\title{Runs through processing steps dynamically}
\usage{
run_process_steps(data, process_options)
}
\arguments{
\item{data}{A data table}

\item{process_options}{list of options for processing}
}
\description{
Runs the processing steps in turn:
calculate_columns_from_existing_data, add_extra_na_cols,
set_negative_values_to_zero. The function constructs the pipe for these
methods dynamically from the argument process_arguments.
}
