% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_countries.R
\name{show_R6_classes}
\alias{show_R6_classes}
\title{Get the countries data is avaliable for}
\usage{
show_R6_classes()
}
\value{
A list of avaliable countries and the region level
data is avaliable for
}
\description{
Show what countries have what level data avaliable.
The function searches the environment for R6 class objects and
extracts the country name and what level it has from the object.
}
