% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{test_cleaning}
\alias{test_cleaning}
\title{Test clean method works correctly}
\usage{
test_cleaning(cntry_obj)
}
\arguments{
\item{cntry_obj}{The DataClass object (R6Class) object to perform checks on.}
}
\description{
Test data can be cleaned correctly using the clean method
}
