% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{test_cleaning}
\alias{test_cleaning}
\title{Expect cleaning runs ok}
\usage{
test_cleaning(self, data_name)
}
\arguments{
\item{self}{The R6Class country class object to perform checks}

\item{data_name}{character_array The name of the class and level to
check}
}
\description{
Test data can be cleaned
}
