% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{test_cleaning}
\alias{test_cleaning}
\title{Test clean method works correctly}
\usage{
test_cleaning(DataClass_obj)
}
\arguments{
\item{DataClass_obj}{The R6Class object to perform checks on.
Must be a \code{DataClass} or \code{DataClass} child object.}
}
\description{
Test data can be cleaned properly. The clean method is invoked
to generate clean data. This data is checked to ensure it is a data.frame,
is not empty, has at least two columns and that columns are clean by calling
\code{expect_clean_cols}. Also tests that \code{avaliable_regions()} are not NA and
they are all characters.
}
\seealso{
Functions used for testing data is cleaned and processed correctly
\code{\link{expect_clean_cols}()},
\code{\link{expect_columns_contain_data}()},
\code{\link{expect_processed_cols}()},
\code{\link{test_download}()},
\code{\link{test_processing}()},
\code{\link{test_return}()}
}
\concept{tests}
