% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{test_download}
\alias{test_download}
\title{Test download method works correctly}
\usage{
test_download(cntry_obj, download, snapshot_path)
}
\arguments{
\item{cntry_obj}{The DataClass object (R6Class) object to perform checks on.}

\item{download}{logical check to download or use a snapshot of the data}

\item{snapshot_path}{character_array the path to save the downloaded
snapshot to.}
}
\description{
Test data can be downloaded if download = TRUE, or a requested
snapshot file is not found, and store a snap shot at the snapshot_path. If an
existing snapshot file is found then just load this data to use in future
tests.
}
