% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{test_download}
\alias{test_download}
\title{Test download method works correctly}
\usage{
test_download(DataClass_obj, download, snapshot_path)
}
\arguments{
\item{DataClass_obj}{The R6Class object to perform checks on.
Must be a \code{DataClass} or \code{DataClass} child object.}

\item{download}{Logical check to download or use a snapshot of the data}

\item{snapshot_path}{character_array the path to save the downloaded
snapshot to.}
}
\description{
Test data can be downloaded if \code{download = TRUE}, or a requested
snapshot file is not found, and store a snap shot in the \code{snapshot_dir}. If
an existing snapshot file is found then load this data to use in future tests
}
\seealso{
Functions used for testing data is cleaned and processed correctly
\code{\link{expect_clean_cols}()},
\code{\link{expect_columns_contain_data}()},
\code{\link{expect_processed_cols}()},
\code{\link{test_cleaning}()},
\code{\link{test_processing}()},
\code{\link{test_return}()}
}
\concept{tests}
