% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{test_processing}
\alias{test_processing}
\title{Test process method works correctly}
\usage{
test_processing(cntry_obj, test_all = FALSE)
}
\arguments{
\item{cntry_obj}{The DataClass object (R6Class) object to perform checks on.}

\item{test_all}{logical. Run tests with all settings (TRUE) or with those
defined in the current class instance (FALSE). Defaults to FALSE.}
}
\description{
Test data can be processed correctly using the process method.
process is invoked to generate processed data which is then checked to ensure
it is a data.frame, which is not empty, has at least 2 columns and calls
\code{expect_processed_columns} to check each column types.
}
