% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{test_processing}
\alias{test_processing}
\title{Test process method works correctly}
\usage{
test_processing(self, data_name, localise = TRUE)
}
\arguments{
\item{self}{The DataClass object (R6Class) object to perform checks on.}

\item{data_name}{character_array The name of the class and level to
check}

\item{localise}{Logical, defaults to TRUE. Should region names be
localised for tests.}
}
\description{
Test data can be processed correctly using the process method
}
