% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{test_processing}
\alias{test_processing}
\title{Expect processing runs ok}
\usage{
test_processing(self, data_name, localise = TRUE)
}
\arguments{
\item{self}{The R6Class country class object to perform checks}

\item{data_name}{character_array The name of the class and level to
check}
}
\description{
Test data can be processed
}
