% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{test_return}
\alias{test_return}
\title{Test return method works correctly}
\usage{
test_return(self, data_name)
}
\arguments{
\item{self}{The DataClass object (R6Class) object to perform checks on.}

\item{data_name}{character_array The name of the class and level to
check}
}
\description{
Test data can be returned correctly using the return method
}
