% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{test_return}
\alias{test_return}
\title{Test return method works correctly}
\usage{
test_return(cntry_obj)
}
\arguments{
\item{cntry_obj}{The DataClass object (R6Class) object to perform checks on.}
}
\description{
Test data can be returned correctly using the return method.
return is invoked to generate returned data which is then checked to ensure
it is a data.frame, not empty and has at least 2 columns. Each column is then
checked to ensure it contains data and is not just composed of NAs.
}
