% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DataClass.R
\name{test_return}
\alias{test_return}
\title{Test return method works correctly}
\usage{
test_return(DataClass_obj)
}
\arguments{
\item{DataClass_obj}{The R6Class object to perform checks on.
Must be a \code{DataClass} or \code{DataClass} child object.}
}
\description{
Test data can be returned correctly using the return method.
return is invoked to generate returned data which is then checked to ensure
it is a data.frame, not empty and has at least 2 columns. Each column is then
checked to ensure it contains data and is not just composed of NAs.
}
\seealso{
Functions used for testing data is cleaned and processed correctly
\code{\link{expect_clean_cols}()},
\code{\link{expect_columns_contain_data}()},
\code{\link{expect_processed_cols}()},
\code{\link{test_cleaning}()},
\code{\link{test_download}()},
\code{\link{test_processing}()}
}
\concept{tests}
