% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{totalise_data}
\alias{totalise_data}
\title{Get totals data given the time series data.}
\usage{
totalise_data(data)
}
\arguments{
\item{data}{A data table}
}
\value{
A data table, totalled up
}
\description{
Get totals data given the time series data.
}
\seealso{
Optional processing function
\code{\link{set_negative_values_to_zero}()}
}
\concept{optional_processing}
